==> 1.data_description.Rmd <==
---
title: Description of data
output: 
  html_document:
    css: mypaper.css
    df_print: kable
    theme: paper
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
---

```{r load, message=FALSE, echo=FALSE}
# set knitr options
source("lib_knitr.R")

# load data
load("data.rda")

# load packages and functions
library("tidyverse")
library("stringr")
library("ggrepel")

source("lib_plot.R")        # colour scales and theme
source("lib_data_manip.R")  # summarise tables
```

## Export database

```{r}
dexport <- select(d, date, n_gear, site, lon, lat, family, genus, species, cpue)
dexport <- dexport %>% mutate_at(vars(lon, lat, cpue), funs(round(., 4)))

# write.table(dexport, file="MEDPLANET_database.csv", quote=F, row.names=F, col.names=T, sep=",")
```

## Spatial scope

Fish larvae were collected at the following sites

```{r}
# wide map
med_coast <- read_csv("data/gshhg_i_2018-03-30_17-54-05.csv", col_types=cols())
map <- ggplot() +  geom_polygon(data=med_coast, aes(x=lon, y=lat), fill="white", color="grey50", size=0.25) +
  annotate("rect", xmin=2.5, xmax=9.9, ymin=41.2, ymax=44, fill=NA, color="grey30", size=1) + coord_quickmap(ylim=c(35, 45))
map
# ggsave("figures/maps_med.pdf", width=6, height=3)

# sampling sites
map + geom_point(data=sites) + geom_text_repel(aes(label=site), data=sites, size=3, segment.alpha=0.7) + coord_quickmap(xlim = c(3, 10))
# ggsave("figures/maps_sampling_sites.pdf", width=6, height=3)
```

## Temporal scope

Sampling occurred over the periods

```{r}
sites_ordered <- c("Port Vendres", "Leucate", "Agde", "Carry", "Marseille", "Cassis", "La Ciotat", "Les Embiez", "Port-Cros", "Villefranche", "Saint Florent", "Bastia", "Bonifacio")
effort$site <- factor(effort$site, levels=sites_ordered)

summary_effort <- effort %>% group_by(site) %>% summarise(begin=min(date), end=max(date, na.rm=T), n_nights_sampled=length(unique(date)), `n_CARE_fished`=sum(n_gear, na.rm=T)) %>% ungroup() #%>% arrange(site)
summary_effort
```

```{r}
# Rename the sites in 'effort' with the sampling number 
effort$site <- factor(effort$site, levels=sites_ordered)

effort$site <- factor(effort$site, levels=levels(effort$site), labels=str_c(summary_effort$site, " (n=", summary_effort$n_CARE_fished, ")"))



ggplot(effort) +
  geom_point(aes(x=date, y=site), shape="|", size = 2, alpha=0.7) +
  scale_y_discrete("",
    limits=rev(levels(effort$site))) +
  theme(axis.title.x=element_blank(), axis.title.y=element_blank())


ggsave("figures/Fig2_sampling_effort.pdf", width=6, height=3.5)
ggsave("figures/Fig2_sampling_effort.png", width=6, height=3.5)
```

Resulting in a total of `r nrow(effort)` nights sampled, the deployment of `r sum(summary_effort$n_CARE_fished)` CAREs, over `r nrow(sites)` sites.

## Sampling effort

At each site, CARE light traps were deployed at several stations. The number of fish caught was summed per site and divided by the number of CARE observed to be working when collected in the morning, to compensate for sampling effort. This Catch Per Unit Effort (cpue = nb fish / CARE / night / site) is the unit used in the rest of the work. Unless otherwise stated, it includes nights with 0 catches.

Statistics on the number of CARE working per night at each site are

```{r}
# effort %>% group_by(site) %>% summarise(mean=round(mean(n_gear, na.rm=T),1), median=median(n_gear, na.rm=T), min=min(n_gear, na.rm=T), max=max(n_gear, na.rm=T)) %>% ungroup() %>% arrange(site)
filter(effort, n_gear>0) %>% group_by(site) %>% sum_tbl("n_gear") %>% select(-n, -n_NA, -starts_with("q"))

```

## Distribution of catches

We focus on *coastal fish* taxa only. A total of `r sum(d$n)` larvae were caught. The most commonly observed catch per night per CARE (CPUE), is 0, by far (the bars represent the number of times a given CPUE was observed in the dataset) and detailed numerical statistics are given below

```{r}
dd <- d0 %>% filter(!is.na(family)) %>% group_by(date, site) %>% summarise(cpue=sum(cpue))
ggplot(dd) + geom_histogram(aes(x=cpue), binwidth=10)

dd %>% ungroup() %>% summarise(min=min(cpue), q25=quantile(cpue, 0.25), med=median(cpue), mean=mean(cpue), q75=quantile(cpue, 0.75), max=max(cpue), sd=sd(cpue), mad=mad(cpue))
#     min   q25   med  mean   q75   max    sd   mad
#   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
# 1     0   0.2 0.778  4.54   2.6  900.  30.3  1.15

# get the number of 0-catch events
round(table(round((d0 %>% group_by(date, site) %>% summarize(sum=sum(cpue)))$sum, 0))/1075, 2)*100
# 0-catch events -> 42% of the nights


```

Excluding 0, pooling all taxa together, and using log10 scaling to see low catches better (NB: notice the labels of the x-axis), the most commonly observed catches are still very low, often around 1 fish/CARE/night/site.

```{r}
ggplot(filter(dd, cpue > 0)) + geom_histogram(aes(x=cpue), binwidth=0.1, na.rm=T) + scale_x_continuous(breaks=c(0, 1, 10, 100, 1000), trans="log10")
```

However, in terms of total number of fish supplied to the coast, its the rare, very high CPUE nights that contribute the most; the following plot is basically the previous one where each bar is multiplied by the number of larvae it corresponds to.

```{r}
ggplot(filter(dd, cpue > 0)) + geom_histogram(aes(x=cpue, weight=cpue), binwidth=0.1, na.rm=T) + scale_x_continuous(breaks=c(0, 1, 10, 100, 1000), trans="log10") + labs(y="number of larvae")
```

```{r}
tot <- sum(dd$cpue)
dd <- dd %>% ungroup() %>% arrange(desc(cpue))
i50 <- min(which(cumsum(dd$cpue) >= tot*0.5))
percent50 <- round(i50 / nrow(dd) * 100)

# exclude nights with particularly high catches
dd <- dd[-c(1,2),]
tot <- sum(dd$cpue)
i50ex <- min(which(cumsum(dd$cpue) >= tot*0.5))
percent50ex <- round(i50ex / nrow(dd) * 100)

i90ex <- min(which(cumsum(dd$cpue) >= tot*0.9))
percent90ex <- round(i90ex / nrow(dd) * 100)
```

Overall, `r nrow(dd)` sampling "events" (one event = one night fished at one site) are recorded in the dataset. Over 50% of the `r sum(d$n)` larvae caught were caught during only `r i50` of those fishing events (`r percent50`% of the effort). Two nights in Corsica resulted in extremely high catches of *S smaris*, which mask everything else. Even excluding those, the `r percent50ex`% richest events represent >50% of the catch and the `r percent90ex`% richest yield >90% (almost all) the catch.

In summary, catches are often *low*, extremely *variable* (probably both geographically and in time) and what matters to coastal populations are the *few nights with high catches*, the rest is just negligible. This variability should be further described spatially and temporally. 

```{r}
# distribution of sampling effort per month
mm <- data.frame(count(unique(select(d0m, date, site, month)), month))
mm

ggplot(mm) + geom_bar(aes(x = month, y = n), stat = "identity", position = "dodge")
```
