==> 2.taxonomic_composition.Rmd <==
---
title: Taxonomic composition
output: 
  html_document:
    css: mypaper.css
    df_print: kable
    theme: paper
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
---

```{r load, message=FALSE, echo=FALSE}
# set knitr options
source("lib_knitr.R")

# load data
load("data.rda")

# load packages and functions
library("plyr")
library("dplyr")
library("stringr")
library("tidyverse")
library("readxl")

source("lib_plot.R")    # colour scales and theme
source("lib_data_manip.R")
```


## Description

### Abundance

Larvae were classified in `r length(unique(na.omit(d$family)))` families, `r length(unique(na.omit(d$genus)))` genera, and `r length(unique(na.omit(d$species)))` identifiable species. 

CPUE is presented (on a log10 scale, beware!) for the most abundant families, genera, and species below (ordered from most abundant to least abundant). Each dot represents the CPUE for one night in one site. The "violin"-shape represents the distribution of observations: the CPUE levels most commonly observed correspond to the wider regions of the shape.

```{r plots_abund, warning=FALSE}
# NB: do not display warnings. All warnings are about log transformation giving -Inf when using 0, which we know

# families
dd <- d0 %>%
  filter(!is.na(family)) %>%
  group_by(date, site, family) %>%
  summarise(cpue=sum(cpue))
dd_stat <- dd %>%
  group_by(family) %>%
  summarise(mean=mean(cpue), median=median(cpue)) %>%
  ungroup() %>% arrange(desc(mean)) %>% head(12)
# TODO check if there are some other interesting families beyond these 12
abundant_families <- dd_stat$family
dd$family <- factor(dd$family, levels=dd_stat$family)
ggplot(na.omit(dd), aes(x=family, y=cpue)) + geom_violin(na.rm=T) + geom_point(alpha=0.2, size=1) + scale_y_continuous(trans="log10") + theme(axis.text.x=element_text(angle=25, hjust=1))

# genera
dd <- d0 %>%
  filter(!is.na(genus)) %>%
  group_by(date, site, family, genus) %>%
  summarise(cpue=sum(cpue))
dd_stat <- dd %>%
  group_by(family, genus) %>%
  summarise(mean=mean(cpue), median=median(cpue)) %>%
  ungroup() %>% arrange(desc(mean)) %>% head(12)
# TODO same as for families
abundant_genera <- dd_stat$genus
dd$genus <- factor(dd$genus, levels=dd_stat$genus)
ggplot(na.omit(dd), aes(x=genus, y=cpue)) + geom_violin(na.rm=T) + geom_point(alpha=0.2, size=1) + scale_y_continuous(trans="log10") + theme(axis.text.x=element_text(angle=25, hjust=1))

# species
dd <- d0 %>%
  filter(!is.na(species)) %>%
  group_by(date, site, family, species, sp) %>%
  summarise(cpue=sum(cpue))
dd_stat <- dd %>%
  group_by(family, species, sp) %>%
  summarise(mean=mean(cpue), median=median(cpue)) %>%
  ungroup() %>% arrange(desc(mean))
dd_stat <- dd_stat %>% head(12)
# TODO same as for families
abundant_species <- dd_stat$species
dd$sp <- factor(dd$sp, levels=dd_stat$sp)
ggplot(na.omit(dd), aes(x=sp, y=cpue)) + geom_violin(na.rm=T) + geom_point(alpha=0.2, size=1) + scale_y_continuous(trans="log10") + theme(axis.text.x=element_text(angle=25, hjust=1))

# add the list of common taxa to what was already in "data.rda"
save(abundant_families, abundant_genera, abundant_species, list=load("data.rda"), file="data.rda")
```

In term of relative total abundance, the dominant taxa are

```{r relative_abund}
#family
dd <- filter(d, family!="Unidentified")
ra_family <- dd %>% group_by(family) %>% summarise(ra = (sum(n)*100)/sum(dd$n)) %>% arrange(desc(ra))
ra_family$cum <- cumsum(ra_family$ra)
head(ra_family, n=10) %>% dplyr::rename(`rel. abund. %`=ra, `cum. abund. %`=cum)

#genus
dd <- filter(d, !is.na(genus))
ra_genus <- dd %>% group_by(genus) %>% summarise(ra = (sum(n)*100)/sum(dd$n)) %>% arrange(desc(ra))
ra_genus$cum <- cumsum(ra_genus$ra)
head(ra_genus, n=10) %>% dplyr::rename(`rel. abund. %`=ra, `cum. abund. %`=cum)

#species
dd <- filter(d, !is.na(species))
ra_species <- dd %>% group_by(species) %>% summarise(ra = (sum(n)*100)/sum(dd$n)) %>% arrange(desc(ra))
ra_species$cum <- cumsum(ra_species$ra)
head(ra_species, n=12) %>% dplyr::rename(`rel. abund. %`=ra, `cum. abund. %`=cum)
```

At the level of a any given taxon, the catch is most often < 1 fish/CARE/night/site and very variable. Despite the extreme variability, the most abundant taxa in CARE catches are not surprising given what we know of the coastal adult populations.


### Distribution

There are `r nrow(sites)` sites. Here we present families, genera, and species that are present in more than 7 sites:

```{r freq_sp}
# compute commonality = number of sites in which a taxon is caught
commonality <- function(d, level) {
  d_common <- d %>%
    filter_(str_c("!is.na(",level,")")) %>%
    group_by_(level) %>%
    summarise(n_sites=length(unique(site))) %>%
    arrange(desc(n_sites))
  d_common[[level]] <- ordered(d_common[[level]])
  return(d_common)
}

# plot taxa in order of commonality
com_families <- commonality(d, "family")
com_families <- filter(com_families, n_sites > 7)
common_families <- as.character(com_families$family)
ggplot(com_families) + geom_bar(aes(x=family, y=n_sites), stat="identity") + theme(axis.text.x=element_text(angle=35, hjust=1))

com_genera <- commonality(d, "genus")
com_genera <- filter(com_genera, n_sites > 7)
common_genera <- as.character(com_genera$genus)
ggplot(com_genera) + geom_bar(aes(x=genus, y=n_sites), stat="identity") + theme(axis.text.x=element_text(angle=35, hjust=1))

com_species <- commonality(d, "species")
com_species <- filter(com_species, n_sites > 7)
common_species <- as.character(com_species$species)
# add abbreviated species
suppressWarnings(com_species <- left_join(com_species, select(taxo, species, sp), by="species"))
com_species$sp <- ordered(com_species$sp)
ggplot(com_species) + geom_bar(aes(x=sp, y=n_sites), stat="identity") + scale_y_continuous(breaks=c(0,5,10)) + theme(axis.text.x=element_text(angle=35, hjust=1))
```

These most common species are similar to the most abundant ones (`r intersect(abundant_species, common_species)` are in common).

`r setdiff(abundant_species, common_species)` are abundant but not common.

`r setdiff(common_species, abundant_species)` are common but not abundant.

```{r save}
# compute the union of abundant and common species
interest_species <- union(abundant_species, common_species)

# describe the interest_species
sum(filter(ra_species, species%in%interest_species)$ra) # 96% of the total catches


# add list of common taxa and species of interest
save(common_families, common_genera, common_species, interest_species, list=load("data.rda"), file="data.rda")
```


## Comparison with adult assemblages

As written above, larvae were classified in `r length(unique(na.omit(d$family)))` families, `r length(unique(na.omit(d$genus)))` genera, and `r length(unique(na.omit(d$species)))` identifiable species.

```{r read_adults, results="hide"}
# read species list of north west mediterranean
library("readxl")
l <- read_excel("data/Liste especes mediterranee.xlsx", sheet=1)

# clean names written in the file
l$family <- str_trim(str_to_lower(l$family))
l$genus <- str_trim(str_to_lower(l$genus))
l$species <- str_trim(str_to_lower(l$species))

# remove undetermined species (1 atherinidae sp.)
l <- filter(l, species != "sp.")

# check that each genus has the same family
filter(summarise(group_by(l, genus), n=length(unique(family))), n>1)

# remove non-demersal fish
non_demersal <- c(
"engraulidae",   # anchois
"scombridae",    # thons...
"clupeidae",     # sardines
"carangidae",    # lichia and trachurus
"sphyraenidae",  # barracudas
"exocoetidae",   # poissons volants
"coryphaenidae", # coryphène
"chlorophthalmidae",
"xiphiidae",     # espadons
"istiophoridae", # marlin
"trichiuridae",  # sabres
"molidae",       # poisson lune
"regalecidae",   # regalec
"trachipteridae",
"macrouridae",   # profond
"aulopidae",     # profond
"petromyzonidae", # eggs in river
"tetraodontidae", # the only mediterranean species is pelagic
"belonidae",     # pelagic
"pomatomidae",   # pelagic
"scaridae"    ,   # because very scarce in France, only one species which come from spain
"batrachoididae" # absent of france
)

l$demersal <- ! l$family %in% non_demersal
l_demersal <- filter(l, demersal == TRUE)


l$family <- str_to_title(l$family)
l$genus <- str_to_title(l$genus)

l_demersal$family <- str_to_title(l_demersal$family)
l_demersal$genus <- str_to_title(l_demersal$genus)

# Join genus and species suffix in species column
l_demersal$species <- str_c(l_demersal$genus, " ", l_demersal$species)
```

According to Louisy:

- Teleost species of North Western Mediterranean are classified in `r length(unique(l$family))` families, `r length(unique(l$genus))` genera, and `r length(unique(l$species))` species. 

- Among which, dermersal species living in French Mediterranean coast represent `r length(unique(l_demersal$family))` families, `r length(unique(l_demersal$genus))` genera, and `r length(unique(l_demersal$species))` species.

Comparing the list in french species recorded in Louisy and the list from CARE catches, the following families of demersal fish were not caught in this study:

`r (uncaught_families <- sort(setdiff(l_demersal$family, d$family)))`.

Among families which are caught, the following genera were never caught either in this study:

```{r uncaught_genera}
filter(l_demersal, ! family %in% uncaught_families &! genus %in% unique(d$genus)) %>% select(family, genus) %>% unique() %>% arrange(family, genus)
```

To summarise:

- `r round((length(unique(na.omit(d$family)))*100)/length(unique(l$family)), 0)`% of families, `r round((length(unique(na.omit(d$genus)))*100)/length(unique(l$genus)), 0)`% of genera, and `r round((length(unique(na.omit(d$species)))*100)/length(unique(l$species)), 0)`% of species recorded in Louisy were caught in CAREs. 

- `r round((length(unique(na.omit(d$family)))*100)/length(unique(l_demersal$family)), 0)`% of families , `r round((length(unique(na.omit(d$genus)))*100)/length(unique(l_demersal$genus)), 0)`% of genera, and `r round((length(unique(na.omit(d$species)))*100)/length(unique(l_demersal$species)), 0)`% of species recorded as french demersal fish were caught in CAREs.
