==> 3.spatial.Rmd <==
---
title: Spatial distribution
output: 
  html_document:
    css: mypaper.css
    df_print: kable
    theme: paper
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
---

```{r load, message=FALSE, echo=FALSE}
# set knitr options
source("lib_knitr.R")

# load data
load("data.rda")

# load packages and functions
# devtools::install_github("jiho/chroma") # CHROMA PACKAGE TO BE DOWNLOADED FROM JIHO's GitHub


library("stringr")
library("lubridate")
library("plyr")
library("tidyverse")
library("reshape2")

library("vegan")
library("FactoMineR")
library("factoextra")
library("NbClust")
library("ggdendro")
library("zoo")
library("RColorBrewer")

source("lib_plot.R")        # colour scales and theme
source("lib_data_manip.R")  # summarise tables
source("lib_factorial.R")   # extract and plot info from CAs
source("lib_quantreg.R")    # perform quantile regression

# Sites ordering
sites_ordered <- c("Port Vendres", "Leucate", "Agde", "Carry", "Marseille", "Cassis", "La Ciotat", "Les Embiez", "Port-Cros", "Villefranche", "Saint Florent", "Bastia", "Bonifacio")
effort$site <- factor(effort$site, levels=sites_ordered)
d0$site <- factor(d0$site, levels=sites_ordered)


#!!!!!!!! IF FOCUSING ALL ANALYSIS ON SUMMER MONTHS ONLY
# Keep only summer months for comparison
d0 <- filter(d0, month%in%c(5:9))
head(d0)

d <- filter(d, month%in%c(5:9))
head(d0)

```

## Abundances among sites

### Total community

Grouping all taxa, the catch per night is highly variable among sites, with some very extreme events yielding very high catches (NB: q25 and q75 are the quantiles at 25% and 75%, the median is the quantile at 50%; those are less sensitive than the mean to extreme values and represent the "general case" better).

```{r summary_table}
dd <- d0 %>% group_by(site, date) %>% summarise(cpue=sum(cpue))

# ggplot(dd) + geom_path(aes(x=cpue, linetype=site, color=site, y=..scaled..), stat="density", bw=0.2) + theme(legend.position=c(0.8, 0.8), legend.key.width=unit(10, "mm")) + labs(y="Scaled probability density")
# last_plot() + scale_x_continuous(trans="identity", limits=c(0,10))

dd_summary <- dd %>% sum_tbl("cpue") %>% select(-n, -n_NA)
dd_summary
```

```{r boxplot}
ggplot(dd) + geom_boxplot(aes(x=site, y=cpue)) + coord_flip()
```

```{r order}
# sort in decreasing order of catches to display stats, below
dd_summary <- arrange(dd_summary, desc(median))
```

Excluding extreme events and focusing on the more common, low (CPUE < 10), catches, allows to highlight that some sites yield better catches (`r head(dd_summary$site, 3)`) and other are consistently low (`r tail(dd_summary$site, 3)`). NB: Keeping only CPUE < 10 only eliminates `r nrow(filter(dd, cpue >= 10))` records out of `r nrow(dd)`. The black squares are quantile=0.9.

```{r zoomed_boxplot, fig.height=4}
# Rename the sites in 'effort' with the sampling number 
summary_effort <- effort %>% group_by(site) %>% summarise(begin=min(date), end=max(date, na.rm=T), n_nights_sampled=length(unique(date)), `n_CARE_fished`=sum(n_gear, na.rm=T)) %>% ungroup()
summary_effort$site <- factor(summary_effort$site, levels=levels(effort$site))
summary_effort
dd$site_effort <- factor(dd$site, levels=levels(dd$site), labels=str_c(levels(dd$site), " (n=", summary_effort$n_CARE_fished, ")"))


ggplot(dd) +
  geom_boxplot(aes(x=site_effort, y=cpue), na.rm=T) +
  geom_quantiled(mapping=aes(x=dd$site_effort, y=dd$cpue), quantiles=0.9, colour="black", size=3, shape=0) +
  scale_x_discrete(limits=rev(levels(dd$site_effort))) + scale_y_continuous("CPUE") +
  coord_flip(ylim=c(0, 18.5)) + theme(axis.title.y=element_blank())
# Save plot
# ggsave("figures/Fig3_CPUE_per_sites.png", width=6, height=4)
# ggsave("figures/Fig3_CPUE_per_sites_SUMMER.pdf", width=6, height=4)

```

```{r map}
# # Same information on a map with q75 in blue, median in green, and q25 in yellow.
# # add site coordinates
# dd_summary <- left_join(dd_summary, sites, by="site")
# cols <- viridis_colors(4)[-1]
# map +
#   geom_point(aes(x=lon, y=lat, size=q75), shape=16, data=dd_summary, colour=cols[1], alpha=0.7) +
#   geom_point(aes(x=lon, y=lat, size=median), shape=16, data=dd_summary, colour=cols[2], alpha=0.7) +
#   geom_point(aes(x=lon, y=lat, size=q25), shape=16, data=dd_summary, colour=cols[3], alpha=0.7) +
#   scale_size_area(max_size=20, guide="none")
```

When tested globally through a quantile based ANOVA, the lower quantiles are surprisingly the most different. The reason is that the higher quantiles (which look very different) may be more difficult to test, hence the lower significance.

```{r among_sites}
aovq(cpue ~ site_effort, tau=c(0.25, 0.5, 0.75, 0.9), data=dd, R=1000)
```

In terms of pairwise differences:

- q=0.25: Saint-Florent is higher than others
- q=0.5: Marseille and Bastia are lower, Villefranche and La Ciotat are higher
- q=0.75: Bastia is lower
- q=0.9: Villefranche is higher than others; probably because it is more easily estimated (more effort).
- wilcox: Marseille and Bastia are lower; La Ciotat, Port-Cros and St-Florent are higher; Carry is often barely signif and would be similar to Marseille.

Also, Saint Florent, Bonifacio, Villefranche, Port-Cros, Les Embiez, La Ciotat, Cassis are never significantly different from each other.

NB: The p-values seem quite sensitive and may change form run to run. We need to reason in terms of global behaviour and trend of the p-value. 

```{r pairwise_among_sites}
# compute pairwise AOV
pwa <- suppressMessages(pairwise.aovq(dd$cpue, dd$site, tau=c(0.25, 0.5, 0.75, 0.9), R=2000))
# and pairwise wilcoxon tests
pww <- suppressWarnings(pairwise.wilcox.test(dd$cpue, dd$site, p.adjust.method="holm"))
pwa$wilcox <- pww
```

```{r plot_among_sites, fig.height=8}
# reformat pairwise p-value tables as data.frame
pw <- ldply(pwa, tidy)
pw$group1 <- factor(pw$group1, levels=sites$site)
pw$group2 <- factor(pw$group2, levels=sites$site)
# label signficant tests
pw$signif <- pw$p.value <= 0.05

# compute the actual values of the quantiles
q <- dd %>% group_by(site) %>% 
  summarise(
    `0.25`=quantile(cpue, 0.25),
    `0.5` =quantile(cpue, 0.5),
    `0.75`=quantile(cpue, 0.75),
    `0.9` =quantile(cpue, 0.9)
  ) %>% 
  gather(key=".id", value="value", -site)
# and the values of the pairwise difference in quantile
pw$q.diff <- q$value[match(pw$group1, q$site)] - q$value[match(pw$group2, q$site)]

# make the matrices symmetric, with difference in quantiles at the top and p-values at the bottom
pw <- bind_rows(
  pw %>% mutate(q.diff=NA),
  dplyr::rename(pw, group1=group2, group2=group1) %>% mutate(p.value=NA)
)

# plot
ggplot(pw, aes(group1, group2)) +
  # plot difference in quantiles
  geom_point(aes(colour=q.diff), na.rm=T, size=5, shape=15) +
  # plot p-value
  geom_point(aes(fill=p.value), na.rm=T, size=6, shape=22, stroke=0) +
  # mark significant differences
  geom_point(colour="red", na.rm=T, size=5, shape=22, data=filter(pw, signif)) +
  # reformat
  facet_wrap(~.id, ncol=2) + coord_fixed() +
  scale_colour_gradient2("Quantile diff\n(row-col)", na.value=NA, low="#7b3294", mid="#f7f7f7", high="#008837") +
  scale_fill_distiller("p-value", palette="PuBu", direction=1, na.value=NA) +
  theme(
    panel.background=element_blank(), panel.grid=element_blank(),
    legend.position=c(0.75, 0.15), legend.direction="horizontal",
    axis.text.x=element_text(angle=65, hjust=1),
    # axis.text.x=element_blank(), axis.ticks.x=element_blank(),
    axis.title=element_blank()
  )
ggsave("figures/pairwise_test_cpue_summer.pdf")

```

Overall, it seems Bonifacio, Cassis, and Carry are difficult to evaluate (low sampling effort); Bastia, Marseille, Agde, Port-Vendres, and Leucate yield low catches; Saint-Florent, Villefranche, Port-Cros, Les Embiez, La Ciotat yield high catches.

Test whether these differences are linked with geomorphological regions:

1. Gulf of Lyon (Port-Vendres -> La Ciotat) vs Ligurian Sea (Les Embiez -> Corsica)
2. East and West of Rhône (but only 3 sites West of Rhône).

These regions are only different regarding high and extreme catches. Fish larvae abundance was higher for sites located in the Ligurian Sea than those located in the Gulf of Lions, and also higher for sites East of Rhône.

```{r abund_per_region}
# add regions for each site
dd <- left_join(dd, sites, by="site")

# test GoL vs Lig
aovq(cpue ~ topography, tau=c(0.25, 0.5, 0.75, 0.9), data=dd)
# SUMMER
# Model 1: cpue ~ topography
# Model 2: cpue ~ 1
#    tau Df Resid Df F value   Pr(>F)    
# 1 0.25  1      758    1.56  0.00561 ** 
# 2 0.50  1      758    9.05 0.000013 ***
# 3 0.75  1      758   27.72 0.000077 ***
# 4 0.90  1      758   79.30  0.00035 ***
  
ggplot(dd) + geom_boxplot(aes(x=topography, y=cpue), na.rm = TRUE) + ylim(0, 10)

# get mean CPUE per topography
dd %>% group_by(topography) %>% summarize(mean=mean(cpue), sd=sd(cpue), q75 = quantile(cpue, probs = 0.75), q90 = quantile(cpue, probs = 0.9))

# SUMMER
#   topography    mean    sd   q75   q90
  # <fct>        <dbl> <dbl> <dbl> <dbl>
# 1 Gulf of Lion  2.41  5.89     2  5.44
# 2 Ligurian sea  8.61 48.7      4 13.9 


# test Rhône
aovq(cpue ~ rhone, tau=c(0.25, 0.5, 0.75, 0.9), data=dd)
# SUMMER
# Model 1: cpue ~ rhone
# Model 2: cpue ~ 1
#    tau Df Resid Df F value   Pr(>F)    
# 1 0.25  1      758    0.06     0.58    
# 2 0.50  1      758    1.63     0.10 .  
# 3 0.75  1      758   19.62 0.000023 ***
# 4 0.90  1      758   86.79 0.000029 ***

ggplot(dd) + geom_boxplot(aes(x=rhone, y=cpue), na.rm = TRUE) + ylim(0, 10)


# get mean CPUE per EW Rhone
dd %>% group_by(rhone) %>% summarize(mean=mean(cpue), sd=sd(cpue), q75 = quantile(cpue, probs = 0.75), q90 = quantile(cpue, probs = 0.9))
```

### Per species

The same tests can be performed per species, to detect if a given species is caught more in the West or in the East. The test was restricted to the most abundant species which are present in both locations. The following table display the mean±SD of the CPUE for each species in each region, together with the statistic and *p*-value of a quantile-based ANOVA between regions.

NB: Test done through rank procedure since bootstrap through anowar does not converge here (=sometimes low sample size, which would make any test problematic).

```{r prepare_sp_data}
# keep only abundant species, but B boops which is not caught in the West
d_abund <- filter(d0, species %in% abundant_species)

# remove the extreme values for S smaris
d_abund <- filter(d_abund, cpue < 200)

# add site characteristics
d_abund <- left_join(d_abund, sites, by="site")
```

```{r sp_topo}
# topography
descriptive_stats <- d_abund %>%
  group_by(sp, topography) %>%
  summarise(stats=str_c(round(mean(cpue),2), "±", round(sd(cpue),1))) %>% 
  spread(key=topography, value=stats)
# perform test per species
tests <- d_abund %>% group_by(sp) %>% do(aovq(cpue ~ topography, tau=c(0.25, 0.5, 0.75, 0.9), data=., test="rank")$table[,c("tau", "pvalue")]) %>% spread(key=tau, value=pvalue)

# make a nice table with both
descriptive_stats <- left_join(descriptive_stats, tests, by="sp") %>% ungroup()
colnames(descriptive_stats) <- c("Species", "CPUE in GoL (mean±SD)", "CPUE in LS (mean±SD)", "q25", "q50", "q75", "q90")
descriptive_stats %>%  mutate(q25=if_else(q25>0.001, q25, 0.001), q50=if_else(q50>0.001, q50, 0.001), q75=if_else(q75>0.001, q75, 0.001), q90=if_else(q90>0.001, q90, 0.001))
#    Species          `CPUE in GoL (mean±SD)` `CPUE in LS (mean±SD)`     q25     q50   q75     q90
#    <chr>            <chr>                   <chr>                    <dbl>   <dbl> <dbl>   <dbl>
#  1 A. hepsetus      0±0.1                   0.12±1.7               0.00627 0.712   0.102 0.00260
#  2 B. boops         0±0                     0.19±1.3               0.0210  0.460   0.355 0.286  
#  3 C. chromis       0.45±3.2                0.78±4.8               0.001   0.0364  0.001 0.001  
#  4 D. annularis     0.15±1.2                0.31±2.2               0.145   0.622   0.619 0.651  
#  5 G. mediterraneus 0.05±0.2                0.04±0.3               0.0210  0.666   0.102 0.001  
#  6 M. surmuletus    0.21±1.1                0.29±2.7               0.00627 0.902   0.619 0.712  
#  7 O. melanura      0±0                     0.26±2.6               0.0143  0.460   0.102 0.0364 
#  8 P. acarne        0.08±0.5                0.13±0.8               0.0143  0.498   0.594 0.0347 
#  9 P. bogaraveo     0.16±1                  0±0                    0.374   0.00682 0.001 0.001  
# 10 P. pilicornis    0.23±1                  0.11±0.9               0.455   0.001   0.001 0.001  
# 11 S. salpa         0.07±0.7                0.14±0.8               0.00954 0.805   0.522 0.203  
# 12 S. smaris        0.29±2.7                0.95±6.7               0.803   0.712   0.102 0.001  

# and the corresponding plot...
ggplot(d_abund, aes(x=sp, y=cpue, colour=topography)) + geom_boxplot(na.rm=T) + scale_y_continuous(trans="log10") + theme(axis.text.x=element_text(angle=35, hjust=1))
```

```{r sp_rhone}
# same for rhone
descriptive_stats <- d_abund %>%
  group_by(sp, rhone) %>%
  summarise(stats=str_c(round(mean(cpue),2), "±", round(sd(cpue),1))) %>% 
  spread(key=rhone, value=stats)
# perform test per species
tests <- d_abund %>% group_by(sp) %>% do(aovq(cpue ~ rhone, tau=c(0.25, 0.5, 0.75, 0.9), data=., test="rank")$table[,c("tau", "pvalue")]) %>% spread(key=tau, value=pvalue)

# make a nice table with both
descriptive_stats <- left_join(descriptive_stats, tests, by="sp") %>% ungroup()
colnames(descriptive_stats) <- c("Species", "CPUE in W (mean±SD)", "CPUE in LS (mean±SD)", "q25", "q50", "q75", "q90")
descriptive_stats %>% mutate(q25=if_else(q25>0.001, q25, 0.001), q50=if_else(q50>0.001, q50, 0.001), q75=if_else(q75>0.001, q75, 0.001), q90=if_else(q90>0.001, q90, 0.001))
#    Species          `CPUE in W (mean±SD)` `CPUE in LS (mean±SD)`     q25     q50     q75     q90
#    <chr>            <chr>                 <chr>                    <dbl>   <dbl>   <dbl>   <dbl>
#  1 A. hepsetus      0±0                   0.09±1.5               0.001   0.001   0.001   0.262  
#  2 B. boops         0±0                   0.15±1.2               0.001   0.001   0.001   0.944  
#  3 C. chromis       0.02±0.1              0.85±4.8               0.00108 0.00166 0.00321 0.001  
#  4 D. annularis     0.02±0.3              0.32±2.1               0.001   0.001   0.115   0.001  
#  5 G. mediterraneus 0.04±0.1              0.05±0.3               0.001   0.001   0.001   0.001  
#  6 M. surmuletus    0.34±1.5              0.23±2.3               0.001   0.001   0.001   0.00142
#  7 O. melanura      0±0                   0.2±2.3                0.001   0.001   0.001   0.591  
#  8 P. acarne        0.12±0.6              0.11±0.7               0.001   0.001   0.001   0.944  
#  9 P. bogaraveo     0.16±1                0.04±0.4               0.001   0.001   0.001   0.001  
# 10 P. pilicornis    0.33±1.2              0.1±0.8                0.001   0.001   0.001   0.001  
# 11 S. salpa         0.1±0.9               0.11±0.7               0.001   0.001   0.001   0.273  
# 12 S. smaris        0±0                   0.89±6.2               0.001   0.001   0.0689  0.001  


# and the corresponding plot...
ggplot(d_abund, aes(x=sp, y=cpue, colour=rhone)) + geom_boxplot(na.rm=T) + scale_y_continuous(trans="log10") + theme(axis.text.x=element_text(angle=35, hjust=1))
```

Overall, many differences are significant when the actual data is mostly noise above 0. Very difficult to interpret.


## Species composition among sites

### Taxonomic richness

Total number of family, genera, and species caught by sampling site.

```{r table_richness, fig.height=5}
dd_sp <- d %>%
  filter(family!="Unidentified") %>%
  group_by(site) %>%
  summarise(
    nb.family = length(unique(na.omit(family))),
    nb.genus = length(unique(na.omit(genus))),
    nb.species = length(unique(na.omit(species)))
  )
dd_sp <- arrange(dd_sp, nb.species)

# add info on effort: number of nights and of CAREs
dd <- d0 %>% group_by(site, date, n_gear) %>% summarise(cpue=sum(cpue)) %>% ungroup()
dd_night <- dd %>% group_by(site) %>% summarise(nb_night=length(unique(date)), nb_care=sum(n_gear)) %>% ungroup()
dd_sp <- left_join(dd_sp, dd_night, by=c("site"))

dd_sp
```

More species are caught in Villefranche probably because sampling is done in winter too.

Species richness vs effort (nb of nights fished or nb of CARE deployed) curves are useful to know if this is just related to differing effort or if there is a true difference in diversity among sites. Here, the curves take a long time to begin to saturate (>150 nights) and never really fully do. 

```{r specaccum}
dd <- d0 %>% group_by(site, date, species) %>% summarise(cpue=sum(cpue)) %>% ungroup()
# cleanup data
dd <- filter(dd, !is.na(species))
dd <- arrange(dd, site, date)

# convert into species matrix
dw <- spread(dd, key="species", value="cpue", fill=0)

# add effort in terms of nb of CAREs (to weight by effort)
dw <- left_join(dw, effort, by=c("site", "date"))

# Species accumulation curves with CI
#
# @param see specaccum
accum_stats <- function(comm, permutations=1000, ...) {
  sp <- specaccum(comm, method="random", permutations=permutations, ...)
  stats <- adply(sp$perm, 1, function(x) {
    data.frame(
      richness=mean(x),
      sd=sd(x),
      ci.low=quantile(x, 0.025, names=FALSE),
      ci.high=quantile(x, 0.975, names=FALSE)
    )
  }, .id="effort")
  if (is.null(sp$effort)) {
    stats$effort <- sp$sites
  } else {
    stats$effort <- sp$effort
  }
  
  return(stats)
}

# compute per night
sp_per_night <- dw %>% group_by(site) %>% do(accum_stats(select(., -site, -date, -n_gear))) %>% dplyr::rename(nights=effort)
# and per CARE
# sp_per_care <- dw %>% group_by(site) %>% do(accum_stats(select(., -site, -date, -n_gear), w=.$n_gear)) %>% rename(`Nb CARE`=effort)
```

```{r plot_specaccum, fig.height=4.5}
# and some plots
differentiating_scales <- list(
  scale_linetype_manual("Sites", values=rep(c("solid", "22", "11"), times=5)),
  scale_fill_discrete("Sites", h=c(0,280)+15), scale_colour_discrete("Sites", h=c(0,280)+15),
  theme(legend.key.width = unit(1, "cm"), legend.position="bottom")
)

# ggplot(sp_per_night) + geom_ribbon(aes(x=nights, ymin=ci.low, ymax=ci.high, fill=site), alpha=0.2) + geom_path(aes(x=nights, y=richness, colour=site, linetype=site))+ differentiating_scales
ggplot(sp_per_night) +
  # geom_vline(aes(xintercept=20) , colour="black", size=0.25) +
  # geom_vline(aes(xintercept=50) , colour="black", size=0.25) +
  # geom_vline(aes(xintercept=150), colour="black", size=0.25) +
  geom_ribbon(aes(x=nights, ymin=richness-sd, ymax=richness+sd, fill=site), alpha=0.2) +
  geom_path(aes(x=nights, y=richness, colour=site, linetype=site)) +
  theme_light() + differentiating_scales +
  labs(x="Nights sampled", y="Species richness")

# Save plot
ggsave("figures/Fig4_species_richness.pdf", width = 6, height = 5.5)
ggsave("figures/Fig4_species_richness_SUMMER.pdf", width = 6, height = 5.5)


# ggplot(sp_per_care) + geom_ribbon(aes(x=`Nb CARE`, ymin=ci.low, ymax=ci.high, fill=site), alpha=0.2) + geom_path(aes(x=`Nb CARE`, y=richness, colour=site, linetype=site)) + differentiating_scales
# ggplot(sp_per_care) + geom_ribbon(aes(x=`Nb CARE`, ymin=richness-sd, ymax=richness+sd, fill=site), alpha=0.2) + geom_path(aes(x=`Nb CARE`, y=richness, colour=site, linetype=site)) + differentiating_scales
```

The richness at an effort of 20 nights (largest where all stations can be tested) show that Carry is significantly less diverse than several other sites. At 50 and 150 nights, Bastia is significantly less diverse than Leucate and Villefranche. When comparable, the others are not significantly different.

```{r compare_specaccum}
# Test difference in species accumulation curves among groups at a given effort level
#
# @param comm community matrix
# @param g factor giving the group for the corresponding element of comm
# @param effort effort level(s) at which to test
# @param nperm number of permutations for the test
# @param p.adjust.method see p.adjust
pairwise.specaccum.test <- function(comm, g, effort=c(20, 50, 150), nperm=1000, p.adjust.method="holm") {
  # force grouping variable (site here) to te a factor (to keep levels consistent)
  g <- as.factor(g)
  ng <- nlevels(g)
  
  # extract richness at given efforts
  richness <- function(x, e) {
    sp <- specaccum(x, method="exact")
    sp$richness[e]
  }

  # reference (i.e. observed) richness values at these levels of effort
  cs <- split(comm, g)
  rich_ref <- laply(cs, richness, e=effort)

  # nperm permutations of sites
  rich_perm <- laply(1:nperm, function(n) {
    cs <- split(comm, sample(g))
    laply(cs, richness, e=effort)
  }, .progress="none")

  # make pairwise tests manually
  # compute all i,j under the diagonal
  ij <- expand.grid(i=1:ng, j=1:ng)
  ij <- filter(ij, i > j)
  # define a function that performs the test for a given couple of i,j indexes (i.e. couple of sites)
  testij <- function(ij) {
    i <- ij[1]
    j <- ij[2]
    # reference difference between sites
    diff_ref <- abs(diff(rich_ref[c(i,j),]))
    # permutation-based distribution of differences between sites
    diff_perm <- aaply(rich_perm[,c(i,j),], 1, function(x) {abs(diff(x))})
    # compute p-value: proportion of time in which differences obtained at random are higher than what is observed
    diffs <- rbind(diff_ref, diff_perm)
    p <- apply(diffs, 2, function(x) {sum(x[-1]>x[1])/(length(x)-1)})
    return(p)
  }
  p <- apply(ij, 1, testij)
  
  # adjust and put those p.values in a easy to display data.frame
  p <- adply(p, 1, function(pp) {
    pp <- p.adjust(pp, method=p.adjust.method, n=sum(!is.na(pp)))
    # NB: removes the NAs to compute the adjusted p.value
    d <- data.frame(ij, p.value=pp)
    return(d)
  })
  p$group1 <- levels(g)[p$i]
  p$group2 <- levels(g)[p$j]
  p$effort <- effort[p$X1]
  p <- select(p, -X1, -i, -j)

  return(p)
}

ps <- pairwise.specaccum.test(comm=select(dw, -date, -site, -n_gear), g=dw$site, effort=c(20, 50, 90))
```

```{r plot_comparison, fig.height=5}
ps$group1 <- factor(ps$group1, levels=sites$site)
ps$group2 <- factor(ps$group2, levels=sites$site)
ps$signif <- ps$p.value < 0.05
# make symmetric
ps <- bind_rows(ps, dplyr::rename(ps, group1=group2, group2=group1))
# plot
ggplot(ps, aes(group1, group2)) +
  geom_point(aes(colour=p.value), na.rm=T, size=4, shape=15) +
  geom_point(colour="red", na.rm=T, size=4, shape=22, data=filter(ps, signif)) +
  facet_wrap(~effort, ncol=2) + coord_fixed() +
  scale_colour_distiller("p", palette="YlGnBu", direction=1, na.value=NA) +
  theme(
    legend.position=c(0.75, 0.15), legend.direction="horizontal",
    axis.text.x=element_text(angle=65, hjust=1),
    # axis.text.x=element_blank(), axis.ticks.x=element_blank(),
    axis.title=element_blank()
  )
ggsave("figures_specaccum_SUMMER.pdf")

```

### Diversity indices by site, region, and topography

```{r diversity, fig.height=4}
# NB: 0 do not matter for the shannon index but influence the number of species in pielou => do not include 0
# per date

# FOR SUMMER
d <- filter(d, month%in%c(5:9))

div <- d %>% group_by(date, site) %>% summarise(shannon = diversity(cpue, index="shannon"), pielou = shannon / log(length(species)) ) %>% ungroup()
# aggregated per month to smooth a bit the immediate variability
div <- d %>% group_by(year, month, site, species) %>% summarise(cpue=mean(cpue)) %>% group_by(year, month, site) %>% summarise(shannon = diversity(cpue, index="shannon"), pielou = shannon / log(length(species)) ) %>% ungroup()

# add site characteristics
div <- left_join(div, select(sites, site, topography, rhone), by="site")

divt <- gather(div, key="index", value="value", shannon, pielou)
ggplot(divt) + geom_boxplot(aes(x=site, y=value, fill = topography), na.rm=T) + facet_wrap(~index, ncol=1, scales="free_y") + theme(axis.text.x=element_text(angle=35, hjust=1))

div_stats <- function(x) {
  with(x,
    data.frame(
      `H mean±sd` = str_c(round(mean(shannon),2), "±", round(sd(shannon),1)), 
      `H med±mad` = str_c(round(median(shannon),2),"±",round(mad(shannon),1)),
      `J mean±sd` = str_c(round(mean(pielou, na.rm=T),2), "±", round(sd(pielou, na.rm=T),1)),
      `J med±mad` = str_c(round(median(pielou, na.rm=T),2),"±",round(mad(pielou, na.rm=T),1)),
    check.names=F)
  )
}
# diversity indices by site and geographic regions
div %>% do(div_stats(.)) %>% as.data.frame()
pp <- div %>% group_by(site) %>% do(div_stats(.)) %>% as.data.frame()
ggplot(reshape2::melt(select(pp, site, `H mean±sd`, `J med±mad`), id.vars = "site"), aes(x = site, y = value)) + geom_point() + facet_wrap(~variable, scales = "free_y", ncol = 1)
div %>% group_by(rhone) %>% do(div_stats(.)) %>% as.data.frame()
div %>% group_by(topography) %>% do(div_stats(.)) %>% as.data.frame()
div %>% group_by(site) %>% do(div_stats(.)) %>% as.data.frame()



#Test differences between site/rhone/topography for each index
# Test for shannon index
wilcox.test(shannon~rhone, div)
wilcox.test(shannon~topography, div)
kruskal.test(shannon~site, div)

# SUMMER: no significant diff

# pairwise.wilcox.test(div$shannon, div$site, p.adjust.method = "holm")


# Test for pielou index
wilcox.test(pielou~rhone, div)
wilcox.test(pielou~topography, div)
kruskal.test(pielou~site, div)

pairwise.wilcox.test(div$pielou, div$site, p.adjust.method = "BH")

# SUMMER 
#KS significant at p=0.05 but pairwise could not identify site
# 	Kruskal-Wallis rank sum test
# 
# data:  pielou by site
# Kruskal-Wallis chi-squared = 21, df = 12, p-value = 0.05

```

### Do the same for the Beta diversity instead

```{r beta diversity, fig.height=4}
# Prepare the community matrix to compute the beta diversity
dbeta <- unique(select(filter(d, !is.na(species)), site, species))
dbeta <- dcast(dbeta, site~species, fill = 0)
head(dbeta)
colsites <- dbeta[,1]
# dbeta <- replace(dbeta[,2:72], dbeta[,2:72]!="0", 1)
# Summer
dbeta <- replace(dbeta[,2:66], dbeta[,2:66]!="0", 1)
dbeta[,] <- sapply(dbeta[,],as.numeric)
rownames(dbeta) <- colsites
head(dbeta)


sites$protection <- c("no-MPA", "MPA", "no-MPA", "MPA", "no-MPA", "no-MPA", "no-MPA", "MPA", "MPA", "no-MPA", "MPA", "MPA", "no-MPA")
data.frame(sites)

m <- betadiver(dbeta, method="w")
River_location <- betadisper(m, group=sites$rhone)
anova(River_location)
plot(River_location)

typo <- betadisper(m, group=sites$topography)
anova(typo)
plot(typo)

reg <- betadisper(m, group=sites$region)
anova(reg)
plot(reg)

mpa <- betadisper(m, group=sites$protection)
anova(mpa)
plot(mpa)

# SUMMER
# no signif diff either but nice differenciation betweet W and E side of the Rhone


```


### Site-species associations

Through a Correspondence Analysis we can study the *relative* proportions of the dominant species at each site. On the plot, sites are close to the species which were most common at this site compared to other species and other sites. In addition, sites are clustered according to a hierarchical clustering based on chi-square distances. The dendrogram cutoff as set through.... 
TODO à revoir.

So overall, the following sites are characterised by a larger proportion of the following species:

- Villefranche: *O melanuara*
- Port-Cros, Agde, and Port-Vendres: *M surmuletus*, *P pilicornis*, *P gattorugine*
- Marseille, Carry: *D annularis*
- Leucate, Bastia: *S salpa*, *P. bogaraveo*, *P acarne*, *S cantharus*
- Bonifacio and Saint Florent: *S smaris*, *A hepsetus* and, in the same group, Les Embiez and Cassis: *C Chromis* and La Ciotat: *G mediterraneus*, *P erythrinus*

Many assocations are geographic but there are some exceptions: Bastia associated with Leucate (probably because of sandy bottom and laguna in both locations); Post-Cros associated with Agde and Port Vendres (no explanation...).

```{r CA_sites, fig.height=5, fig.width=8}
dd <- d0 %>% filter(species %in% interest_species) %>% group_by(site, species) %>% summarise(mean=mean(cpue))
# dd <- d0 %>% filter(species %in% c(interest_species, "Aidablennius sphynx", "Diplodus puntazzo", "Diplodus sargus", "Epinephelus marginatus", "Gymnammodytes cicerelus", "Lypophrys trigloides", "Lithognathus mormyrus", "Parablennius rouxi", "Parablennius sanguinolentus", "Serranus cabrilla")) %>% group_by(site, species) %>% summarise(mean=mean(cpue))
ddm <- as.multivar(dd, "site")


# perform hierachical clustering
dd_stand <- decostand(ddm, method = "chi.square") # Chi2 transform the data
dist_matrix <- vegdist(dd_stand, "euclidian") # compute distance on transformed data
hdw <- hclust(dist_matrix, method="ward.D2") # compute hierarchical clustering
plot(hdw)

# Color dendrogram with groups
cut <- 7 # ward.D2 cut at 5, but cutting at 6 enables to discriminate Vlfr from Corsica on Axe 4
dendr <- dendro_data(hdw, type = "rectangle") 
clust <- cutree(hdw, k = cut)               # find 'cut' clusters
clust.df <- data.frame(label = names(clust), cluster = clust)

# Split dendrogram into upper grey section and lower coloured section
height <- unique(dendr$segments$y)[order(unique(dendr$segments$y), decreasing = TRUE)]
cut.height <- mean(c(height[cut], height[cut-1]))
dendr$segments$line <- ifelse(dendr$segments$y == dendr$segments$yend &
   dendr$segments$y > cut.height, 1, 2)
dendr$segments$line <- ifelse(dendr$segments$yend  > cut.height, 1, dendr$segments$line)

# Number the clusters
dendr$segments$cluster <- c(-1, diff(dendr$segments$line))
change <- which(dendr$segments$cluster == 1)
for (i in 1:cut) dendr$segments$cluster[change[i]] = i + 1
dendr$segments$cluster <-  ifelse(dendr$segments$line == 1, 1, 
             ifelse(dendr$segments$cluster == 0, NA, dendr$segments$cluster))
dendr$segments$cluster <- na.locf(dendr$segments$cluster) 

# Consistent numbering between segment$cluster and label$cluster
clust.df$label <- factor(clust.df$label, levels = levels(dendr$labels$label))
clust.df <- arrange(clust.df, label)
clust.df$cluster <- factor((clust.df$cluster), levels = unique(clust.df$cluster), labels = (1:cut) + 1)
dendr[["labels"]] <- merge(dendr[["labels"]], clust.df, by = "label")

# Positions for cluster labels
n.rle <- rle(dendr$segments$cluster)
N <- cumsum(n.rle$lengths)
N <- N[seq(1, length(N), 2)] + 1
N.df <- dendr$segments[N, ]
N.df$cluster <- N.df$cluster - 1

# Plot the dendrogram
cols <- colorRampPalette(brewer.pal(8, "Paired"))
cols <- cols(8)
myPal <- c( "grey60", cols)
# myPal <- c("grey60", "#053061", "#3288bd", "#92c5de", "#d53e4f","#7fbc41", "#fdae61")

ggplot() + 
   geom_segment(data = segment(dendr), 
      aes(x=x, y=y, xend=xend, yend=yend, size=factor(line), colour=factor(cluster)), 
      lineend = "square", show.legend = FALSE) + 
   scale_colour_manual(values = myPal) +
   scale_size_manual(values = c(.5, 1)) +
   geom_text(data = N.df, aes(x = x, y = y, label = factor(cluster),  colour = factor(cluster + 1)), 
      hjust = 1.5, show.legend = FALSE) +
   geom_text(data = label(dendr), aes(x, y, label = label, colour = factor(cluster)), 
       hjust = -0.2, size = 4, show.legend = FALSE) +
   geom_hline(aes(yintercept=2.3), linetype=2)+
   scale_y_reverse(expand = c(0.2, 0)) + 
   labs(x = NULL, y = NULL) +
   coord_flip() +
    theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        panel.background = element_rect(fill = "white"),
        panel.grid = element_blank())



# Save plot
ggsave("figures/dendrogram_sites_SUMMER.pdf", width = 10, height = 3)


# Try to find cutoff with some criterion
NbClust(data=ddm, diss=dist_matrix, distance=NULL, min.nc=2, max.nc=9, method="ward.D2", index=c("kl", "ch", "cindex", "db", "silhouette", "duda", "pseudot2", "beale", "gap"))
# 3 clusters seems to be the best cutoff, but Vlfr with Bastia and Leucate --> due to higher sampling effort=higher diversity ?


# perform CA
ca <- CA(ddm, graph=F)

# assemble all data
ca_df <- augment(ca)
# part <- data.frame(group=factor(cutree(hdw, k=6)))
# part$label <- rownames(part)

# Order to match dendrogram colors
part <- arrange(dendr$labels, cluster)
part$cluster <- as.numeric(as.character(part$cluster))-1
ca_df <- left_join(ca_df, select(part, label, group=cluster), by=c("label"))
ca_df$group <- factor(ca_df$group)


# compute variance explained
eig <- tidy(ca)
dims <- names(ca_df)[str_detect(names(ca_df), "Dim")]
labs <- str_c(dims, " (",round(ca$eig[1:4,2]),"%)")

ca_df <- arrange(ca_df, group)
# ca_df$label[which(ca_df$type=="col")] <- paste("italic('", ca_df$label[which(ca_df$type=="col")], "')", sep="")
# ca_df$label[which(ca_df$type=="Port Vendres")] <- "Port-Vendres"
# ca_df$label[which(ca_df$type=="La Ciotat")] <- "La_Ciotat"
# ca_df$label[which(ca_df$type=="Les Embiez")] <- "Les_Embiez"
# ca_df$label[which(ca_df$type=="Saint Florent")] <- "Saint-Florent"

# plot result
p1 <- ggplot(ca_df, aes(x=Dim1, y=Dim2, colour=group)) +
  scale_colour_manual(guide="none", values = myPal[2:9], na.value="grey50") +
  geom_point(size=1) +
  ggrepel::geom_text_repel(aes(label=label), size=4.2, segment.alpha=0.8) +
  scale_x_continuous(breaks=0) + scale_y_continuous(breaks=0) +
  labs(x=labs[1], y=labs[2]) + theme_light(base_size=14)
p2 <- ggplot(ca_df, aes(x=Dim3, y=Dim4, colour=group)) +
  scale_colour_manual(guide="none", values = myPal[2:9], na.value="grey50") +
  geom_point(size=1) +
  ggrepel::geom_text_repel(aes(label=label), size=4.2, segment.alpha=0.8) +
  scale_x_continuous(breaks=0) + scale_y_continuous(breaks=0) +
  labs(x=labs[3], y=labs[4]) + theme_light(base_size=14)

# plot and save graph
pdf(file = "figures/CA_species_sites_SUMMER.pdf", width = 10, height = 8)
glayout(list(p1, 1, 1), list(p2, 1, 2))
dev.off()
```

Using a one-way PERMANOVA based on Chi-2 distance we can explicitely test for the significance of the difference of communities between geographic or topographic regions. Species assemblages were not different between site located in the Ligurian Sea and those located in the Gulf of Lions (`topograhy`); they were according to which side of the Rhône they were located on.

```{r}
ddm$rhone <- sites$rhone
ddm$topography <- sites$topography
adonis(dist_matrix~rhone+topography, ddm, permutations = 999)$aov.tab
# Permutation: free
# Number of permutations: 999
# 
# Terms added sequentially (first to last)
# 
#            Df SumsOfSqs MeanSqs F.Model    R2 Pr(>F)  
# rhone       1      11.3   11.27    2.56 0.181   0.02 *
# topography  1       7.0    7.02    1.59 0.113   0.14  
# Residuals  10      44.1    4.41         0.707         
# Total      12      62.4                 1.000         
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

# SUMMER: same results, higher signif
# Number of permutations: 999
# 
# Terms added sequentially (first to last)
# 
#            Df SumsOfSqs MeanSqs F.Model    R2 Pr(>F)   
# rhone       1      15.4   15.37    3.80 0.251  0.003 **
# topography  1       5.5    5.48    1.36 0.089  0.211   
# Residuals  10      40.4    4.04         0.660          
# Total      12      61.3                 1.000          
```


## Synchronicity among sites

### Time of arrival of abundant species among sites

```{r time_of_arrival, fig.height=8}
# mean CPUE per week, site, and species for abundant species
dd <- d0 %>% filter(species %in% abundant_species) %>% group_by(yweek, year, site, species) %>% summarise(cpue=mean(cpue))
# we choose a granularity of one week to smooth out day-today variability and still keep resolution

# create a new weekly date coordinate
dd$date <- ymd(str_c(dd$year, "-01-01")) + weeks(dd$yweek)
# and a new fake one starting all the same year
dd$fake_date <- ymd("2010-01-01") + weeks(dd$yweek)

# rescale abundances per species, site, and year, to get a comparable peak per year
dd <- dd %>% group_by(species, site, year) %>% mutate(scaled_cpue=cpue/max(cpue))
dd$scaled_cpue[is.na(dd$scaled_cpue)] <- 0  # when max = 0

# rename species for display
dd$species_abbr <- abbrev_sp(dd$species, n=4)

# plot
ggplot(dd) +
  geom_line(aes(x=fake_date, y=scaled_cpue, colour=site, linetype=site)) +
  facet_grid(species_abbr~year)+
  scale_x_date("", date_breaks="2 month", date_labels="%b", minor_breaks=NULL) +
  theme(
    axis.text.x=element_text(angle=45, hjust=1),
    legend.position="bottom"
  ) +
  scale_linetype_manual(values=rep(c("solid", "22", "11"), times=5)) +
  scale_colour_discrete(h=c(0,280)+15) +
  no_y
```

Difficult to read...

### Anomaly in date of arrival among sites

The time of arrival of each species throughout the year is plotted to detect its peak. Winter species are shifted by 183 days (half a year).

```{r compute_anomaly, fig.height=9}
# Add species caught several times and at several sites + interest_species
dd <- d %>% 
  filter(species %in% c(interest_species, "Aidablennius sphynx", "Diplodus puntazzo", "Diplodus sargus", "Epinephelus marginatus", "Gymnammodytes cicerelus", "Lypophrys trigloides", "Lithognathus mormyrus", "Parablennius rouxi", "Parablennius sanguinolentus", "Serranus cabrilla")) %>%
  group_by(date, year, yday, site, species) %>% summarise(cpue=sum(cpue)) %>% ungroup()

# for winter species, shift the time
dd$yday[dd$species %in% c("Sarpa salpa", "Pagellus acarne")] <- (dd$yday[dd$species %in% c("Sarpa salpa", "Pagellus acarne")] + 183) %% 365

# compute the mean arrival day by species/sites/year
dp <- dd %>% filter(cpue > 0) %>% group_by(species, site, year) %>% summarise(peak=weighted.mean(yday, cpue, na.rm=TRUE))
# compute reference yday for each species
dp <- dp %>% group_by(species) %>% mutate(mean_peak = mean(peak))
dp$lag <- dp$peak - dp$mean_peak


# check that the detected reference yday makes sense
ggplot(dd) + facet_wrap(~species, scales="free_y", ncol=4) + theme(legend.position="bottom") +
  geom_point(aes(x=yday, y=cpue, colour=site), alpha=0.5) + scale_y_log10() +
  geom_vline(aes(xintercept=mean_peak), data=unique(select(dp, species, mean_peak)))
```

This leads to remove *Boops boops* which is caught in only two sites and *Sarpa salpa* which has a bimodal pattern. The the time lag between the reference arrival date (line) and the observed arrival date at each site are investigated per site.

Port Vendres is late. Eastern sites are mostly early (La Ciota, Villfranche, and Bastia significantly so). But Port Cros is significantly late, hence breaking the pattern.

```{r test_anomaly, fig.height=5}
dp <- filter(dp, ! species %in% c("Boops boops", "Sarpa salpa"))
ggplot(filter(dp, ! site == "Carry")) +
  geom_vline(xintercept=0, size=1, colour="grey70") +
  #geom_histogram(aes(x=lag), binwidth=1) +
  geom_density(aes(x=lag, y=..count..), adjust=0.5, fill='white') +
  facet_wrap(~site, ncol=3) + theme_light(base_size = 14) + 
  scale_x_continuous("Lag (in days)") + scale_y_continuous("Density", limits=c(0, 1.3), expand=c(0,0))

# Save figure
# ggsave("figures/Fig9_spatial_lag_sites_density.pdf", width = 8, height = 7)

t <- dp %>%
  group_by(site) %>%
  do({
    data.frame(
      mean=mean(.$lag, na.rm=T),
      sd=sd(.$lag, na.rm=T),
      med=median(.$lag, na.rm=T),
      mad=mad(.$lag, na.rm=T),
      glance(wilcox.test(.$lag, mu = 0, alternative = "two.sided")) %>% select(statistic, p.value)
    )
  }) %>% ungroup()
t$signif <- stars(t$p.value)
t
# # A tibble: 13 × 8
#             site   mean    sd    med   mad statistic p.value signif
#           <fctr>  <dbl> <dbl>  <dbl> <dbl>     <dbl>   <dbl>  <chr>
# 1        Leucate   4.32    31   6.12    26       547  0.3785       
# 2   Port Vendres  17.17    35  18.83    24       120  0.0052     **
# 3           Agde   9.72    32   9.48    22       112  0.2645       
# 4          Carry  -5.05    25  -7.63    32         5  0.6250       
# 5      Marseille   1.52    40  -4.63    28       319  0.4639       
# 6         Cassis -14.91    22  -8.02    11         4  0.1094       
# 7      La Ciotat -16.78    39 -14.62    24        37  0.0013     **
# 8     Les Embiez  -1.56    38  -8.05    19       172  0.1406       
# 9      Port-Cros  17.40    35  11.41    25       285  0.0200      *
# 10  Villefranche  -8.53    26  -6.86    16       390  0.0171      *
# 11     Bonifacio   0.99    22   0.49    27       107  0.9563       
# 12 Saint Florent   8.59    28   2.90    13       329  0.2312       
# 13        Bastia -11.69    40  -6.39    22       246  0.0269      *
```


```{r test_anomaly_timing_among_basins, fig.height=9}


# Add species caught several times and at several sites + interest_species
dd <- d %>% 
  filter(species %in% c(interest_species)) %>%
  group_by(date, year, yday, site, species) %>% summarise(cpue=sum(cpue)) %>% ungroup()

# for winter species, shift the time
dd$yday[dd$species %in% c("Sarpa salpa", "Pagellus acarne")] <- (dd$yday[dd$species %in% c("Sarpa salpa", "Pagellus acarne")] + 183) %% 365

# Add site characteristics
dd <- left_join(dd, select(sites, site, topography, rhone), by="site")

# compute the mean arrival day by species/sites/year
dp <- dd %>% filter(cpue > 0) %>% group_by(species, rhone, year) %>% summarise(peak=weighted.mean(yday, cpue, na.rm=TRUE))
# compute reference yday for each species
dp <- dp %>% group_by(species) %>% mutate(mean_peak = mean(peak))
dp$lag <- dp$peak - dp$mean_peak

# check that the detected reference yday makes sense
ggplot(dd) + facet_wrap(~species, scales="free_y", ncol=4) + theme(legend.position="bottom") +
  geom_point(aes(x=yday, y=cpue, colour=rhone), alpha=0.5) + scale_y_log10() +
  geom_vline(aes(xintercept=mean_peak), data=unique(select(dp, species, mean_peak)))

# Remove unrelevant spp
dp <- filter(dp, ! species %in% c("Boops boops", "Sarpa salpa", "Atherina hepsetus", "Oblada melanura", "Spicara smaris"))

# Plot the lag
ggplot(dp) +
  geom_vline(xintercept=0, size=2, colour="grey70") +
  geom_histogram(aes(x=lag), binwidth=1) +
  geom_density(aes(x=lag, y=..count..)) +
  facet_wrap(~rhone, ncol=3) + theme_light(base_size = 14) + 
  scale_x_continuous("Lag (in days)") + scale_y_continuous("Count (lag/species/year)")

# Compute test for significativity for East and West of Rhone separately (each w/r to 0)
t <- dp %>%
  group_by(rhone) %>%
  do({
    data.frame(
      mean=mean(.$lag, na.rm=T),
      sd=sd(.$lag, na.rm=T),
      med=median(.$lag, na.rm=T),
      mad=mad(.$lag, na.rm=T),
      glance(wilcox.test(.$lag, mu = 0, alternative = "two.sided")) %>% select(statistic, p.value)
    )
  }) %>% ungroup()
t$signif <- stars(t$p.value)
t
# # A tibble: 2 × 8
#           rhone  mean    sd   med   mad statistic p.value signif
#          <fctr> <dbl> <dbl> <dbl> <dbl>     <dbl>   <dbl>  <chr>
# 1 West of Rhone   9.1    28   8.6    17       409   0.057      .
# 2 East of Rhone  -6.5    21  -7.9    21       346   0.012      *

# Test for differences between East and West of Rhone directly
glance(wilcox.test(filter(dp, rhone=="West of Rhone")$lag, filter(dp, rhone=="East of Rhone")$lag, alternative = "g")) %>% select(statistic, p.value)
#   statistic p.value
# 1      1130  0.0028

# SUMMER: now non significant results

# Compare by Topography

# compute the mean arrival day by species/sites/year
dp <- dd %>% filter(cpue > 0) %>% group_by(species, topography, year) %>% summarise(peak=weighted.mean(yday, cpue, na.rm=TRUE))
# compute reference yday for each species
dp <- dp %>% group_by(species) %>% mutate(mean_peak = mean(peak))
dp$lag <- dp$peak - dp$mean_peak

# Remove unrelevant spp
dp <- filter(dp, ! species %in% c("Boops boops", "Sarpa salpa", "Atherina hepsetus", "Oblada melanura", "Spicara smaris"))

t <- dp %>%
  group_by(topography) %>%
  do({
    data.frame(
      mean=mean(.$lag, na.rm=T),
      sd=sd(.$lag, na.rm=T),
      med=median(.$lag, na.rm=T),
      mad=mad(.$lag, na.rm=T),
      glance(wilcox.test(.$lag, mu = 0, alternative = "two.sided")) %>% select(statistic, p.value)
    )
  }) %>% ungroup()
t$signif <- stars(t$p.value)
t
# # A tibble: 2 × 8
#     topography  mean    sd   med   mad statistic p.value signif
#         <fctr> <dbl> <dbl> <dbl> <dbl>     <dbl>   <dbl>  <chr>
# 1 Gulf of Lion   3.7    27   5.4    24       613    0.61       
# 2 Ligurian sea  -3.7    23  -6.5    19       430    0.16 


# Test for differences between Topography directly
glance(wilcox.test(filter(dp, topography=="Gulf of Lion")$lag, filter(dp, topography=="Ligurian sea")$lag, alternative = "two.sided")) %>% select(statistic, p.value)
#   statistic p.value
# 1      1263    0.23


# SUMMER same results

```

```{r emblematic species}
# Mérou
data.frame(filter(d, species == "Epinephelus marginatus"))

# Anguille
data.frame(filter(d, species == "Anguilla anguilla"))
sum(data.frame(filter(d, species == "Anguilla anguilla"))$n) # 45 

# Dorades
data.frame(filter(d, species == "Sparus aurata"))
sum(data.frame(filter(d, species == "Sparus aurata"))$n)


data.frame(filter(d, species == "Spondyliosoma cantharus"))

# Denti
data.frame(filter(d, species == "Dentex dentex"))

# Corb
data.frame(filter(d, species == "Sciaena umbra"))
```
