==> 4.temporal.Rmd <==
---
title: Temporal variability
output: 
  html_document:
    css: mypaper.css
    df_print: kable
    theme: paper
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
---

```{r load, message=FALSE, echo=FALSE}
# set knitr options
source("lib_knitr.R")

# load data
load("data.rda")

# load packages and functions
library("reshape2")
library("plyr")
library("dplyr")
library("tidyverse")
library("stringr")
library("FactoMineR")
library("printr")
library("lubridate")
library("doParallel")
registerDoParallel(cores=min(detectCores()-1, 12))

source("lib_plot.R")         # colour scales and theme
source("lib_data_manip.R")   # summarise tables
source("lib_factorial.R")    # extract and plot info from CAs
source("lib_quantreg.R")     # perform quantile regression
source("lib_quantilogram.R") # compute quantilogram

q <- c(.25, .5, .75, .9)    # quantiles of interest
```


## Interannual variability

```{r interannual}
dd <- d0 %>% group_by(date, site, year) %>% summarize(cpue=sum(cpue)) %>% ungroup()
dd$days <- as.numeric((dd$date - min(dd$date)) / (24*3600))
dd

dd %>% group_by(year) %>% summarize(mean=mean(cpue))

ggplot(dd) + geom_violin(aes(x=year, y=cpue)) + geom_point(aes(x=year, y=cpue), alpha=0.2) + scale_y_continuous("CPUE", trans='log', breaks=c(1, 100)) + scale_x_discrete("") + geom_point(aes(x=year, y=median(cpue)), shape=3, size=10)
```

## Main patterns

The time series of the total CPUE (all species pooled) at each site highlights a seasonal variability, with two peaks of high catches around June and in Autumn.

```{r ts_data}
dd <- d0 %>% group_by(site, date, year) %>% summarize(cpue=sum(cpue)) %>% ungroup()
dd$days <- as.numeric((dd$date - min(dd$date)) / (24*3600))
```

```{r srq}
# NB: the linear fit, below, is better because it highlights the autumn peaks better.

# system.time(srq_fit <- srq(dd$days, dd$cpue, tau=q, interval="confidence", df=30, se="boot", R=1000))
# 
# srq_fit$date <- min(dd$date) + srq_fit$x * 24 * 3600
# srq_fit$lower[srq_fit$lower < 0] <- 0
# 
# ggplot(srq_fit) +
#   geom_point(aes(x=date, y=cpue), data=filter(dd, cpue < 100), size=1, alpha=0.5, shape=16) +
#   geom_ribbon(aes(x=date, ymin=lower, ymax=higher, fill=tau, alpha=tau)) +
#   scale_alpha_manual(values=c(0., 0.2, 0., 0.2)) +
#   geom_line(aes(x=date, y=fit,colour=tau)) +
#   geom_rug(aes(x=date), filter(dd, cpue > 100), sides="t", alpha=0.5) +
#   geom_rug(aes(x=date), dd, sides="b", alpha=0.1)
```

```{r llrq}
# system.time(llrq_fit <- llrq(dd$days, dd$cpue, tau=q, bw=20, n=200, .parallel=T, .progress="text", smooth=1, interval="conf", se="nid"))
system.time(llrq_fit <- llrq(dd$days, dd$cpue, tau=q, bw=20, n=200, .parallel=F, .progress="text", smooth=1, interval="conf", se="boot", bsmethod="xy", R=1000))

llrq_fit$date <- min(dd$date) + llrq_fit$x * 24 * 3600
llrq_fit$lower[llrq_fit$lower < 0] <- 0

head(llrq_fit)
llrq_fit <- rename(llrq_fit, Tau = tau)

# same plot with different data
ggplot(llrq_fit) +
  geom_point(aes(x=date, y=cpue), data=filter(dd, cpue < 100), size=1, alpha=0.5, shape=16) +
  geom_ribbon(aes(x=date, ymin=lower, ymax=higher, fill=Tau, alpha=Tau)) +
  scale_alpha_manual(values=c(0.2, 0.2, 0.2, 0.2)) +
  geom_line(aes(x=date, y=fit,colour=Tau)) +
  geom_rug(aes(x=date), filter(dd, cpue > 100), sides="t", alpha=0.5) +
  geom_rug(aes(x=date), dd, sides="b", alpha=0.1) + theme_light(base_size = 14) + scale_x_datetime("Sampling dates") + scale_y_continuous("CPUE")

# Save plot
# ggsave("figures/Fig6_timeseries.pdf", width = 10, height = 6)
```

TODO do the same per genus/species

In terms of composition, three main assemblages can be distinguished:
- Winter, with *P acarne* et *S salpa*,
- Spring and early summer, with a few main species quite spread out (associated with different months),
- Summer and early fall, with high diversity, and a similar composition among months.

```{r CA, fig.height=5}
dd <- d0 %>% filter(species %in% interest_species) %>% group_by(month, species) %>% summarise(mean=mean(cpue))

dd$month <- as.factor(month.abb[dd$month])
ca <- CA(as.multivar(dd, "month"), graph=F)
autoplot(ca, size=1, dimensions=c(1,2)) + scale_color_manual("", label = c("Species", "Month"), values = c("grey40", "#d53e4f"), guide = "none") + theme_light(base_size = 9)

# print eigvalues and var explained
ca$eig

ggsave("figures/CA_month.pdf", width = 4, height = 5)

```


## Dominant time scales of variability

Abundances seem to vary at various scales: seasonal but maybe also lunar, etc. We also want to know if abundances are correlated from one day to the next. We can inspect the dominant scales of variability through a correlogram, which tests correlation at various scales. Here we use a quantilogram, which focuses on variability of data around various quantiles rather than around the mean, like a regular correlogram does.


We start by computing the quantilogram for the total community (all species pooled) for the three sites where sampling is performed at high enough frequency for the quantilogram to be defined.

```{r total_quantilogram}
# focus on sites where the quantilogram can be well defined (other sites do not have enough data at all)
dd <- d0 %>% filter(site %in% c("Leucate", "Villefranche", "Bastia")) %>% group_by(site, date) %>% summarise(cpue=sum(cpue)) %>% ungroup()

# time coordinates and breaks
dd$days <- as.numeric((dd$date - min(dd$date)) / (24*3600))
breaks <- c(1:7, seq(7, 31, by=2), seq(31, 365*1.2, by=5))

# compute quantilogram and conf int using permutations
qm <- dd %>% group_by(site) %>% do({
  # reference
  qm <- quantilogram(.$days, .$cpue, dist.max=365*1.2, breaks=breaks, tau=q) %>% tidy()

  # permutations
  pqm <- laply(1:1000, function(i) {
    as.matrix(quantilogram(.$days, sample(.$cpue), dist.max=365*1.2, breaks=breaks, tau=q)$quantilogram)
  }, .parallel=F)
  qm$q1 <- melt(aaply(pqm, 2:3, quantile, probs=0.025))$value
  qm$q2 <- melt(aaply(pqm, 2:3, quantile, probs=0.975))$value
  
  qm
})

qm <- rename(qm, Tau=tau)

# Add ribbon to day 0
qm$q1[which(qm$dx==0)] <- 0
qm$q2[which(qm$dx==0)] <- 0


# plot result
sx <- list(scale_x_continuous("Lag", breaks=c(0, 1, 3, 7, 29.5, 182, 365), minor_breaks=seq(0, 356*1.2, by=30.5), labels=c(0, "1 day", "3 days", "1 week", "1 lunar\nmonth", "6 months", "1 year"), trans='log1p'), scale_y_continuous("Autocorrelation"))
ggplot(qm) + facet_grid(Tau~site) +
  geom_linerange(aes(x=dx, ymin=0, ymax=corr, color=Tau), size=1.2) +
  # geom_hline(aes(yintercept=sel * qnorm(0.05/2) * c(-1,1)), alpha=0.5, colour="red") +
  geom_ribbon(aes(x=dx, ymin=q1, ymax=q2, fill = Tau), alpha=0.3, data=filter(qm, dx >= 0)) + sx + coord_cartesian(ylim=c(-0.5, 1), xlim=c(0, 450)) + theme_light(base_size = 14) + theme(axis.text.x=element_text(hjust=1, angle=45))# + geom_vline(aes(xintercept = 80))

# save graph
ggsave("figures/Fig9_quantilograms.pdf", width = 10, height = 7)
ggsave("figures/Fig9_quantilograms.png", width = 10, height = 7)


# and zoom on small scale variability
last_plot() + coord_cartesian(ylim=c(-0.5, 1), xlim=c(0, 80))

# save plot
# ggsave("figures/FigS3b_quantilograms.pdf", width = 10, height = 6)

```

The confidence interval in Bastia clearly shows that the sampling was centred around the new moon which makes it difficult to estimate what happens at invervals that are not multiple of a lunar month.

Still Bastia seems to highlight a periodicity of ~5.5 months for mid to high captures, which the others do not. No idea what that can be. In all three, but in Villefranche expecially, there seems to be high correlation at a bit less than a lunar cycle period. Finally, all highlight high correlation at low scales for low and mid capture ranges. This is likely driven by successive zeros. This short scale correlation is not apparent for high captures, highlighting that high capture events are isolated and exceptional.

Even for the most abundant species, we do not have enough data to compute the same quantilograms per species.

## Lunar phase variations

```{r results="hide"}
suppressPackageStartupMessages(library("oce", quietly=TRUE))
# get all dates
dates <- sort(unique(d0$date))

# for each date, compute the proportion of the night when the moon is up
moon <- ldply(dates, function(x) {
  # x <- dates[8]

  # go back from noon the previous day, to noon on the day of collection
  interval <- 0.1
  xr <- x - days(1) + seq(12, 36, by=interval)*3600

  # compute position of sun and moon
  sun_alt <- sunAngle(xr, lon=7.31, lat=43.68)$altitude
  moon <- moonAngle(xr, lon=7.31, lat=43.68)
  moon_alt <- moon$altitude

  # detect when sun is down and moon is up
  moon_visible <- (sun_alt < 0) & (moon_alt > 0)
  # compute the number of hours this is the case
  moon_time <- sum(moon_visible) * interval
  
  # compute a moon phase index, based on the proportion illuminated
  # 0 : new moon
  # 1 : next full moon
  # -1 : previous full moon
  # negative : descending
  # positive : ascending
  moon_illum <- mean(moon$illuminatedFraction[sun_alt < 0]) * sign(diff(moon$illuminatedFraction[c(1, length(xr))])) * moon_time
  
  # Compute days since / from new moon
  moon_phase <- mean(moon$phase[sun_alt < 0])%%1
  if(moon_phase < 0.5) {
  moon_phase <- moon_phase + 1
  }
  orbital_period <- 27.32  # orbital period of the moon, in days
  moon_phase <- - round(28 - (moon_phase * orbital_period))

  
    # compute average "moonlight" indexes during that time
  if (moon_time > 0) {
    # visible fraction
    moon_index1 <- mean(moon$illuminatedFraction[moon_visible]) / moon_time
  
    # fraction * diameter * altitude
    moon_index2 <- sum(moon$illuminatedFraction[moon_visible] * moon$diameter[moon_visible] * (moon$altitude[moon_visible]/90))
  } else {
    moon_index1 <- moon_index2 <- 0
  }
  
  data.frame(date=x, moon_time, moon_illum, moon_phase, moon_index1, moon_index2)
}, .progress="none")

# add that to the data
d0m <- left_join(d0, moon, by="date")
dm <- left_join(d, moon, by="date")

# check distribution of effort per moonphase
count(d0m, moon_phase)
#  -14  -13  -12  -11  -10   -9   -8   -7   -6   -5   -4   -3   -2   -1    0    1    2    3    4    5    6    7    8    9   10   11   12   13 
# 1600  900 1000 1300 1300 1600 1900 2400 2200 2800 3400 4401 5700 8402 9900 8800 9400 7700 6600 5500 4100 4400 2900 2200 1900 2500 1800  900 

# Plot effort at each moon phase
ggplot(count(d0m, moon_phase)) + geom_bar(aes(x = abs(moon_phase), y = n), stat = "identity") + scale_x_continuous("Lunar day")
# ggsave("figures/effort_moonphase.png")


# # interpolate times of rise and set of celestial bodies
# rise_set <- function(time, altitude) {
#   # time <- ymd("2016-01-01") + hours(0:(24*3))
#   # altitude <- sunAngle(time, lon=7.31, lat=43.68)$altitude
#   xn <- as.numeric(time)
#
#   # detect sign changes in altitude
#   sign_change <- diff(sign(altitude))
#   i <- which(sign_change != 0)
#
#   # interpolate time when body crosses altitude 0
#   d <- data.frame(x1=xn[i], x2=xn[i+1], y1=altitude[i], y2=altitude[i+1])
#   times <- apply(d, 1, function(X) {approx(X[3:4], X[1:2], value)$y})
#   times <- as.POSIXct(times, tz=tz(time), origin="1970-01-01")
#
#   # determine crossing direction
#   direction <- c("set", "rise")[as.numeric(sign_change[i]>0)+1]
#
#   return(data.frame(time=times, direction))
# }
```

If we look at the total number of larvae caught, a large proportion of it is caught around the new moon, as shown below for common species. However the second plot shows that there is still considerable variability in CPUE, even around the new moon, and that it mostly masks the moon-phase signal (the red lines are polynomial regressions of the average CPUE; most of which do not highlight a moon phase influence).

```{r fig.height=5}

# average per lunar day
dplot <- filter(d0m,  cpue < 200) %>% group_by(species, moon_phase) %>% summarize(cpue = mean(cpue))

moon_x <- list(scale_x_continuous(breaks=c(-14, 0, 14), labels=c("\U0099", "\U0098", "\U0099")), theme(axis.text.x=element_text(family="Wingdings 2"), axis.title.x=element_blank()))

# scale between 0 and 1
range01 <- function(x){(x-min(x))/(max(x)-min(x))}
dplot <- ddply(dplot, ~species, function(x){
  x$cpue <- range01(x$cpue)
  x
})

# compute average arrival 
rangeMoon <- dplot %>% group_by(species) %>% summarise(range = weighted.mean(moon_phase, w = cpue))
dplot <- left_join(dplot, rangeMoon, by = "species")
dplot <- arrange(dplot, range)

# order species per mean of arrival during lunar month
# unique(dplot$species)
# dplot$species <- factor(x = dplot$species, labels = unique(dplot$species)[c(7, 16, 10, 2:3, 5:6, 8:9, 11:15, 18:19, 4, 17, 1)], levels = unique(dplot$species)[c(7, 16, 10, 2:3, 5:6, 8:9, 11:15, 18:19, 4, 17, 1)])
dplot$species <- factor(x = dplot$species, labels = unique(dplot$species), levels = unique(dplot$species), exclude=NULL)

# plot distribution of arrival through lunar month
ggplot(filter(dplot, !species %in% "Atherina hepsetus") ) + geom_bar(aes(x=moon_phase, y=cpue), fill="grey30", colour=NA, stat = 'identity') + facet_wrap(~species, ncol=3, scales = "free_y") + moon_x + labs(y="total number of larvae (scaled per species)")


# # density
# ggplot(filter(d0m, species %in% abundant_species)) + geom_density(aes(x=moon_phase, weight=cpue, y=..scaled..), bw=0.05, fill="grey30", colour=NA) + facet_wrap(~species, ncol=3) + moon_x + no_y
#
# # density logged
# ggplot(filter(d0m, species %in% abundant_species)) + geom_density(aes(x=moon_phase, weight=log1p(cpue), y=..scaled..), bw=0.05, fill="grey30", colour=NA) + facet_wrap(~species, ncol=3) + moon_x + no_y
# # does not change much

# ggplot(filter(d0m, species %in% abundant_species), aes(x=moon_phase, y=cpue)) + geom_point(size=1, alpha=0.5) + geom_smooth(se=F, colour="#F55D5B", size=0.5, data=filter(dm, species %in% abundant_species), method="loess") + facet_wrap(~species, ncol=3, scales="free_y") + moon_x + no_y + labs(y="cpue (scaled per species)")

# # log10 transformed
# ggplot(filter(d0m, species %in% abundant_species), aes(x=moon_phase, y=cpue)) + geom_point(size=1, alpha=0.5) + geom_smooth(se=F, colour="red", size=0.5) + facet_wrap(~species, ncol=3) + scale_y_log10() + moon_x
```

Even when we reduce the data to summer months (May to August, included), during which the catches are better overall, the moon phase signal is still mostly indiscernible in instantaneous CPUE.

```{r fig.height=5}
# reduce to summer months
in_summer <- function(x) {
  month(x) %in% 5:8
}
ggplot(filter(d0m, species %in% abundant_species, in_summer(date)), aes(x=moon_phase, y=cpue)) + geom_point(size=1, alpha=0.5) + geom_smooth(se=F, colour="#F55D5B", size=0.5, data=filter(dm, species %in% abundant_species), method="loess") + facet_wrap(~species, ncol=3, scales="free_y") + moon_x + no_y + labs(y="summer cpue (scaled per species)")
```



Another way to look at it is to regress CPUE on moon phase. Using quantile regression, we can focus on the common scenario (median, quantile=0.5), or the higher catches (quantiles > 0.5). We see that the median catch is basically 0 everywhere and that only the 90th percentile (quantile=0.9) starts to respond to moon phase.



```{r fig.height=4}
suppressPackageStartupMessages(library("quantreg", quietly=TRUE))

# Check effort in summer only
cc <- data.frame(count(unique(select(filter(d0m, in_summer(date)), date, site, moon_phase)), moon_phase))
# Plot effort at each moon phase
ggplot(cc) + geom_bar(aes(x = abs(moon_phase), y = n, fill = 1/n), stat = "identity", position = "dodge")


# Check effort throughout the sampling period
cc <- data.frame(count(unique(select(d0m, date, site, moon_phase)), moon_phase))
# Plot effort at each moon phase
ggplot(cc) + geom_bar(aes(x = abs(moon_phase), y = n), stat = "identity", position = "dodge") + scale_x_continuous("abs(lunar day)") + scale_y_continuous("Sampling effort")
# The effort is (slightly) better distributed over all moon phase when considering the entire period
ggsave("figures/effort_moonphase.png", height=6, width=8)

# Join d0m with fishing effort of each moon_phase
dd <- left_join(d0m, rename(cc, n_nights=n))

# select data: common species, in summer, without extreme Spicara
dd <- filter(dd, species %in% interest_species, !species %in% "Atherina hepsetus", cpue < 250) %>% group_by(species, moon_phase) %>% summarize(cpue = mean(cpue))
# Or all data (except massive events) but that inflates the number of 0s for rare species
# dd <- filter(dd, cpue < 250) %>% group_by(species, moon_phase) %>% summarize(cpue = weighted.mean(cpue, 1/n_nights))


# Fn to standardize cpue per species (each species gets the same weight)
range01 <- function(x){(x-min(x))/(max(x)-min(x))}
dd <- ddply(dd, ~species, function(x){
  x$cpue <- range01(x$cpue)
  x
})


# only consider distance to the new moon, irrespective of direction
dd$moon_phase <- abs(dd$moon_phase)


# # plot according to various moon related variables
# ggplot(dd) + geom_point(aes(x=moon_time, y=cpue))
# ggplot(dd) + geom_point(aes(x=moon_phase, y=cpue))
# ggplot(dd) + geom_point(aes(x=moon_index1, y=cpue))
# ggplot(dd) + geom_point(aes(x=moon_index2, y=cpue))


# Compute rql for all quantiles
quantiles <- c(0.25, 0.5, 0.75, 0.9)
rm(qpred, qpreds)
qpred <- rql(cpue ~ moon_phase, tau=quantiles, data=dd)

# check significance of the regression curves
summary.rql(qpred, se="boot", R=1000, bsmethod="xy")


# new data for prediction
x <- seq(0, 14, length.out=14)
new <- data.frame(moon_phase=x)
qpreds <- predict.rql(qpred, new = new, interval="confidence")
qpreds$moon_phase <- rep(x, times = length(unique(quantiles)))
# Convert tau to factor 
qpreds$quantile <- factor(qpreds$tau)


# If all regressions are not predictable (e.g. all O)
# x <- seq(0, 14, length.out=15)
# new <- data.frame(moon_phase=x)
# qpreds <- predict.rql(qpred[3:4], new = new, interval="confidence")
# qpreds$moon_phase <- rep(x, times = 2)


# Keep only uniques 
dd$cpue_r <- round(dd$cpue, 2) 
  
dplot <- dd %>% select(-species) %>% group_by(moon_phase) %>% count(cpue_r) %>% rename(Occurrence=n) %>% ungroup() %>% as.data.frame()

# plot the result
ggplot(mapping=aes(x=moon_phase)) +
  geom_ribbon(aes(ymin=lower, ymax=higher, fill=quantile), data=qpreds, alpha=0.3) +
  geom_path(aes(y=fit, colour=quantile), data=qpreds) + 
  geom_jitter(aes(y=cpue), data=dd, alpha=0.5, height=0, width=0.5) +
  scale_color_discrete("Quantiles", labels = c(expression("25%"^"***"), expression("50%"^"***"), expression("75%"^"***"), expression("90%"^"***"))) + 
  scale_fill_discrete("Quantiles", labels = c(expression("25%"^"***"), expression("50%"^"***"), expression("75%"^"***"), expression("90%"^"***"))) +
  # scale_colour_brewer(palette="Reds") + scale_fill_brewer(palette="Reds") +
  scale_y_continuous("CPUE (scaled per species)") + coord_cartesian(ylim = c(0.02, 0.97)) +
  scale_x_continuous("Moon phase", breaks=c(0, 7, 14), labels=c("\U0098", "\U00BA", "\U0099"), limits=c(0, 14), oob = scales::squish) +
  theme_light(base_size = 14) + theme(axis.text.x=element_text(family="Wingdings 2"),legend.text.align = 0)

ggsave("figures/Fig8_Moon_influence_jitter.png", width = 8, height = 5)
#
```

So, overall, we observe a significant influence of the moon, with higher catches around the new moon for all quantiles, the influence being higher for higher quantiles : highest catches systematically occur around the new moon, even more than average catches. 


## Seasonal variations

Unsurprisingly, the catch is very seasonal, with higher catches from May to mid-September and another small increase around November. NB: the two points with very high catches of *S smaris* occurred in June; they were removed from the plot.

```{r}
dd <- d0 %>% group_by(date, yday, year, site) %>% summarise(cpue=sum(cpue))
dd$fake_date <- ymd("2012-01-01") + days(dd$yday) # put 2012 because it had 366 days

# Plot yearly patterns of CPUE per site
ggplot(dd) + geom_point(aes(x=fake_date, y=cpue)) + geom_smooth(aes(x=fake_date, y=cpue)) + facet_wrap(~site, scales="free_y")


q <- c(.25, .5, .75, .9)

# srq_fit <- srq(dd$fake_date, dd$cpue, tau=q, interval="confidence", df=30, se="boot", R=1000)
# 
# srq_fit$fake_date <- srq_fit$x
# srq_fit$lower[srq_fit$lower < 0] <- 0
# 
# ggplot(srq_fit) +
#   geom_point(aes(x=fake_date, y=cpue), data=filter(dd, cpue < 100), size=1, alpha=0.5, shape=16) +
#   geom_ribbon(aes(x=fake_date, ymin=lower, ymax=higher, fill=tau, alpha=tau)) +
#   scale_alpha_manual(values=c(0., 0.2, 0., 0.2)) +
#   geom_line(aes(x=fake_date, y=fit,colour=tau)) +
#   geom_rug(aes(x=fake_date), filter(dd, cpue > 100), sides="t", alpha=0.5) +
#   geom_rug(aes(x=fake_date), dd, sides="b", alpha=0.1) + scale_x_date("month", date_breaks="1 month", date_labels="%b")
```

```{r llrq2}
# registerDoParallel(cores=2)
# system.time(llrq_fit <- llrq(dd$fake_date, dd$cpue, tau=q, bw=20, n=200, .parallel=T, .progress="text", smooth=1, interval="conf", se="nid"))
system.time(llrq_fit <- llrq(dd$yday, dd$cpue, tau=q, bw=20, n=200, .parallel=T, .progress="text", smooth=1, interval="conf", se="boot", bsmethod="xy", R=1000))
# stopImplicitCluster()

llrq_fit$fake_date  <- ymd("2012-01-01") + llrq_fit$x
llrq_fit$lower[llrq_fit$lower < 0] <- 0

head(llrq_fit)
llrq_fit <- rename(llrq_fit, Tau = tau)

# same plot with different data
ggplot(llrq_fit) +
  geom_point(aes(x=fake_date, y=cpue), data=filter(dd, cpue < 30), size=1, alpha=0.5, shape=16) +
  geom_ribbon(aes(x=fake_date, ymin=lower, ymax=higher, fill=Tau, alpha=Tau)) +
  scale_alpha_manual("Quantiles", label = c("25%", "50%", "75%", "90%"), values=c(0.2, 0.2, 0.2, 0.2)) +
  geom_line(aes(x=fake_date, y=fit,colour=Tau)) +
  geom_rug(aes(x=fake_date), filter(dd, cpue >= 30), sides="t", alpha=0.5) +
  geom_rug(aes(x=fake_date), dd, sides="b", alpha=0.1) + 
  scale_x_date("Month",  date_breaks="1 month", labels = c("Dec","Jan", "Feb", "Mar","Apr", "May","Jun","Jul","Aug","Sep","Oct","Nov","Dec","Jan","Feb")) + theme_light() + 
  scale_color_discrete("Quantiles", label = c("25%", "50%", "75%", "90%")) + 
  scale_fill_discrete("Quantiles", label = c("25%", "50%", "75%", "90%")) + 
  scale_y_continuous("CPUE") 


# Save plot
# ggsave("figures/Fig9a_seasonality_total_catches.pdf", width = 7, height = 4)


# Find min and max of q90
head(llrq_fit)
llrq_fit %>% group_by(Tau) %>% summarize(min = min(fit), max=max(fit), sd = sd(fit))


##--- Repeat but for each site separately

# Loop over sites to create each plot
plots <- dlply(dd, ~site, function(x){
  
  system.time(llrq_fit <- llrq(x$yday, x$cpue, tau=q, bw=20, n=200, .parallel=T, .progress="text", smooth=1, interval="conf", se="boot", bsmethod="xy", R=1000))
  # stopImplicitCluster()
  
  llrq_fit$fake_date  <- ymd("2012-01-01") + llrq_fit$x
  llrq_fit$lower[llrq_fit$lower < 0] <- 0
  
  head(llrq_fit)
  llrq_fit <- rename(llrq_fit, Tau = tau)
  
  # same plot with different data
  p <- ggplot(llrq_fit) +
    geom_point(aes(x=fake_date, y=cpue), data=x, size=1, alpha=0.5, shape=16) +
    geom_ribbon(aes(x=fake_date, ymin=lower, ymax=higher, fill=Tau, alpha=Tau)) +
    scale_alpha_manual(guide="none", "Quantiles", label = c("25%", "50%", "75%", "90%"), values=c(0.2, 0.2, 0.2, 0.2)) +
    geom_line(aes(x=fake_date, y=fit,colour=Tau)) +
    # geom_rug(aes(x=fake_date), filter(x, cpue >= 30), sides="t", alpha=0.5) +
    geom_rug(aes(x=fake_date), x, sides="b", alpha=0.1) + 
    scale_x_date("Month",  date_breaks="1 month", labels = c("Dec","Jan", "Feb", "Mar","Apr", "May","Jun","Jul","Aug","Sep","Oct","Nov","Dec"," ","Feb"), limits=c(ymd("2012-01-01"), ymd("2012-12-31"))) + theme_light() + 
    scale_color_discrete(guide="none", "Quantiles", label = c("25%", "50%", "75%", "90%")) + 
    scale_fill_discrete(guide="none", "Quantiles", label = c("25%", "50%", "75%", "90%")) +
    scale_y_continuous("CPUE") + ggtitle(paste(unique(x$site)))
  
  return(p)
  
})

# plot all together
pdf(file = "figures/yearly_CPUE_all_sites.pdf", width=12, height=12)
grid.arrange(plots$Leucate, plots$`Port Vendres`, plots$Agde, plots$Carry, plots$Marseille, plots$Cassis, plots$`La Ciotat`, plots$`Les Embiez`, plots$`Port-Cros`, plots$Villefranche, plots$Bonifacio, plots$`Saint Florent`, plots$Bastia, ncol=3)
# ggsave("figures/yearly_CPUE_all_sites.pdf")
dev.off()

#--- Check species richness over the season
head(d)
div <- d %>% group_by(month) %>% summarise(div = length(unique(species)))
ggplot(div, aes(x = month, y = div)) + geom_bar(stat = "identity")


```


## Synchronicity among years

Time arrival anomaly => see 3.Spatial.Rmd

At species level, the time of arrival has consistent seasonality across years but can vary of a few weeks from year to year.

```{r fig.height=9}
# dd <- d0 %>% filter(year != 2016, species %in% abundant_species) %>% group_by(yweek, year, species) %>% summarise(cpue=mean(cpue))
# 
# # create a fake date coordinate
# dd$fake_date <- ymd("2014-01-01") + weeks(dd$yweek)
# 
# # rescale per species and year
# dd <- dd %>% group_by(species, year) %>% mutate(scaled_cpue=cpue/max(cpue))
# dd$scaled_cpue[is.na(dd$scaled_cpue)] <- 0  # when max = 0
# 
# # order species by average/first date of arrival
# order <- dd %>%
#   group_by(species) %>%
#   filter(cpue > median(cpue)) %>% 
#   summarise(avg_date=weighted.mean(yweek, scaled_cpue), first_date=min(yweek)) %>%
#   arrange(first_date)
# 
# # rename species for display
# order$species_abbr <- abbrev_sp(order$species, n=5)
# 
# # force order by time of arrival
# dd$species_abbr <- factor(dd$species, levels=order$species, labels=order$species_abbr)
# 
# ggplot(dd) +
#   geom_path(aes(x=fake_date, y=scaled_cpue, colour=year)) +
#   facet_grid(species_abbr~.) +
#   scale_x_date(date_breaks="1 month", date_labels="%b", minor_breaks=NULL) +
#   labs(y="cpue (scaled per species and year)") + no_y


# Do it with points instead of path
dd <- d0 %>% filter(year != 2016) %>% group_by(yweek, year, species) %>% summarise(ycpue=mean(cpue)) %>% group_by(yweek, species) %>% summarize(cpue = mean(ycpue)) # average CPUE per week and per year among site, then average per week within year to get a single value per week

# rescale per species
dd <- dd %>% group_by(species) %>% mutate(scaled_cpue=cpue/max(cpue))
dd$scaled_cpue[is.na(dd$scaled_cpue)] <- 0  # when max = 0

# order species by average/first date of arrival
order <- dd %>%
  group_by(species) %>%
  filter(cpue > median(cpue)) %>%
  summarise(avg_date=weighted.mean(yweek, scaled_cpue), first_date=min(yweek)) %>%
  arrange(first_date)

# rename species for display
order$species_abbr <- abbrev_sp(order$species, n=5)
order <- filter(order, !species_abbr == ". ")

# force order by time of arrival
dd$species_abbr <- factor(dd$species, levels=rev(c(order$species[c(5, 7:length(order$species))], order$species[c(1, 4, 2, 3, 6)])), labels=rev(c(order$species_abbr[c(5, 7:length(order$species_abbr))], order$species_abbr[c(1, 4, 2, 3, 6)])))

# Change cpue = 0 to NA 
dd$scaled_cpue[which(dd$scaled_cpue == 0)] <- NA

#  Remove unidentified species
dd <- filter(dd, !is.na(species))

# Check if a single value per week
head(dd)
filter(dd, species == "Pagellus acarne")

# plots
ggplot(dd) +
  geom_point(aes(x=yweek, y=species_abbr, size = scaled_cpue), alpha=0.7) +
  # scale_x_date(date_breaks="1 month", date_labels="%b", minor_breaks=NULL) +
  labs(y="CPUE (scaled per species)") + scale_size_area("CPUE", max_size = 6) + scale_x_continuous("", breaks = seq(1, 48, length = 12), labels = c("Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")) + theme(axis.text.y = element_text(face="italic")) + theme_light() + scale_y_discrete(expand = c(0,1))

# save plot
# ggsave("figures/Fig9b_weekly_abund_species.pdf", width = 7, height = 9)



```


## Abundance among years

The effort varied quite a lot among years.

```{r}
effort %>% mutate(year=lubridate::year(date)) %>% filter(year != 2016) %>% group_by(year) %>% summarise(n_nights_sampled=length(unique(date)), n_CARE_fished=sum(n_gear))
```

In total CPUE, pooling all taxa together, the catch seems similar over the years, with possibly a shift towards higher catches in 2014 and 2015. 

```{r}
dd <- d0 %>% filter(year != 2016, year != 2012) %>% group_by(date, year, site) %>% summarise(cpue=sum(cpue))

ggplot(dd, aes(x=year, y=cpue)) + geom_violin(na.rm=T) + geom_point(alpha=0.2, size=1) + scale_y_continuous(trans="log10")

# Or boxplot, easier to check for differences
ggplot(dd, aes(x=year, y=cpue)) + geom_boxplot(na.rm=T) + geom_point(alpha=0.2, size=1) + scale_y_continuous(trans="log10")

```

Pairwise aovq test pvalues
```{r warning=FALSE}

dd <- d0 %>% filter(!year == 2016, !year == 2012) %>% group_by(date, year, site) %>% summarise(cpue=sum(cpue))
dd$year <- factor(dd$year, levels = c(2013, 2014, 2015))


test <- aovq(cpue ~ year, tau=c(0.25, 0.5, 0.75, 0.9), data=dd)

pw <- suppressMessages(pairwise.aovq(dd$cpue, dd$year, tau=c(0.25, 0.5, 0.75, 0.9), R=1000))
# plot
(pw <- ldply(pw, function(x) {
  # x <- pw$`0.9`
  x <- reshape2::melt(x$p.value)
  x$Var1 <- factor(x$Var1, levels=levels(dd$year))
  x$Var2 <- factor(x$Var2, levels=levels(dd$year))
  x$signif <- round(x$value, 2) <= 0.05
  # x <- data.frame(x)
  return(x)
}))

```


```{r}
ggplot(pw) + geom_point(aes(Var1, Var2, shape=signif, colour=signif), na.rm=T) + theme(axis.text.x = element_text(angle=30, hjust=1), axis.title=element_blank()) + facet_wrap(~.id, ncol=1)
```


## Species composition among years

When inspecting the dynamics per species the characteristic year-species couples are:

- 2012: *A hepsetus*
- 2013: *S smaris*
- 2014: *B boops*, *D annularis*, *O melanuara*

```{r fig.height=5}
dd <- d0 %>% filter(year != 2016, year != 2012, species %in% interest_species) %>% group_by(year, species) %>% summarise(mean=mean(cpue))
# Summer only
dd <- d0 %>% filter(year != 2016, year != 2012, species %in% interest_species, month%in%c(5:9)) %>% group_by(year, species) %>% summarise(mean=mean(cpue))


# relative abundance
# ggplot(dd) + geom_bar(aes(x=year, y=mean), stat="identity") + facet_wrap(~species, scales="free_y") + theme(axis.text.x=element_text(angle=30, hjust=1))

# just make the plot cleaner by pre-computing the relative abundances
ddr <- dd %>% group_by(species) %>% mutate(mean=mean/max(mean))
ggplot(ddr) +
  geom_bar(aes(x=year, y=mean), stat="identity", width=0.75) +
  facet_wrap(~species, ncol=3) +
  scale_y_continuous("relative cpue", breaks=c(0, 0.5, 1)) +
  theme(axis.text.x=element_text(angle=30, hjust=1), axis.text.y=element_blank(), axis.ticks.y=element_blank())

ca <- CA(as.multivar(dd, x.var="year"), graph=F)
autoplot(ca) + scale_color_manual("", label = c("Species", "Year"), values = c("grey40", "#d53e4f"), guide = "none") + theme_light(base_size = 9)
ca$eig

ggsave("figures/FigS2_CA_years_SUMMER.pdf", width = 5, height = 6)

```


Differences in species composition are observed between years. 2012, 2013, 2014 are characterized by some species (2012 = *A hepsetus*, 2013 = *S smaris*, 2014 = *B boops*, *D annularis*, *O melanuara*).

These differences could be due to a different sampling effort. For example, 2012 is characterized by A.hepsetus maybe because sites sampled in 2012 have essentially *A.hepsetus*.

However, years 2013 2014 have the same (or very similar) sampling effort (in terms of sites sampled).

2013 : All sites except La Ciotat, Cassis, Carry, Les Embiez
2014 : All sites except La Ciotat, Cassis
2015 : Only Embiez, Ciotat, Cassis, Marseille, Villefranche.

So difference between years in species composition is not due to difference in sampling sites. 


## Synchonicity among years 

Time arrival per year

```{r, fig.height=6}
# sum catches per day and site

# Keep only Vlfr that has been sampled all three years
dd <- d %>% filter(species %in% interest_species, site%in%c("Villefranche")) %>% group_by(date, year, yday, yweek, site, species) %>% summarise(cpue=sum(cpue)) %>% ungroup()
unique(dd$site)

# Remove zeros
dd <- filter(dd, cpue!=0)

## Check time of arrival by species to identify winter species
ggplot(dd) +
  geom_histogram(aes(x=yweek, fill=year), binwidth=1) +
  #geom_histogram(aes(x=yweek, weight=cpue),binwidth=1) +
  geom_density(aes(x=yweek, y=..count..)) +
  facet_wrap(~species, ncol=3) + geom_vline(xintercept = 0,linetype="dashed") + xlab("")



###--- Winter species should be handle differently (because among two year)

# Select winter species 
winter <- rbind(dd[which(dd$species=="Sarpa salpa"),], dd[which(dd$species=="Pagellus acarne"),])

# Switch date by 1/2 year
fake_date <- yday("2012-07-01")
fake_week <- week("2012-07-01")


## Modify day and week according to this new date
winter$yday_shifted <- round(yday(winter$date)-fake_date, 0) %% 365 # diff of number of day since mid year, and modulo 365 (days per year) to center on new year -> min=1,  max=183


# Same for week
winter$yweek_shifted <- round(week(winter$date)-fake_week, 0) %% 52
# same as for days, but max=26


## So if we remove the season 2015-2016 (because of no data in the beggining of 2016) we can
## consider 2012-2013 = 2012 season, 2013-2014 = 2013 season, 2014-2015 = 2014 season
winter$year_shifted <- winter$year
winter$year_shifted[which(winter$year == 2012 & winter$yday > 183)] <- "2013"
winter$year_shifted[which(winter$year == 2013 & winter$yday > 183)] <- "2014"
winter$year_shifted[which(winter$year == 2014 & winter$yday > 183)] <- "2015"

# Prepare df like for summer species
winter <- select(winter, date, year=year_shifted, yday=yday_shifted, yweek=yweek_shifted, site, species, cpue)

```

2012 = only species which recruit in summer  
2013 = winter species (season 2012-2013) + summer species of 2013  
2014 = winter species (season 2013-2014) + summer species of 2014  
2015 = winter species (season 2015-2016) + summer species of 2015

```{r}

# Select summer species
summer <- dd %>% filter(species %in% interest_species) %>% filter(!species %in% c("Sarpa salpa", "Pagellus acarne")) %>% ungroup()

# Add winter species
samples <- rbind(summer, winter)

head(samples)
unique(samples$species)
#Compute the mean arrival week by species/sites/year
ref_detailed <- samples %>% group_by(species,site,year) %>% summarise(ref_week_det=round(weighted.mean(yweek, cpue),0))
ref_sp <- ref_detailed %>% group_by(species) %>% summarise(ref_week_sp = round(mean(ref_week_det),0))

#Join the two data.frame to comute lag time for each site
lag <- left_join(ref_detailed, ref_sp, by=c("species"))
lag$lag_week <- lag$ref_week_det - lag$ref_week_sp


#  Remove 2012 because not full
lag <- filter(lag, !year == 2012)


ggplot(lag) +
  geom_histogram(aes(x=lag_week),binwidth=1) +
  geom_density(aes(x=lag_week, y=..count..)) +
  facet_wrap(~year, ncol=1)+geom_vline(xintercept = 0,linetype="dashed") + xlab("")


```

```{r, fig.height=4}
#Compute mean and median lag
descriptive_stats <- lag %>%
  group_by(year) %>%
  summarise(stats=str_c(round(mean(lag_week),2), "±", round(sd(lag_week),1)), median=median(lag_week))

suppressWarnings (test <- lag %>% group_by(year) %>% do(glance(wilcox.test(.$lag_week,mu = 0,alternative = "two.sided"))))

# make a nice table with both
left_join(descriptive_stats, select(test, statistic, p.value), by="year") %>% ungroup()
#   year  stats     median statistic p.value
#   <fct> <chr>      <dbl>     <dbl>   <dbl>
# 1 2013  2.17±7.5     1       104.   0.207 
# 2 2014  -0.89±5     -1.5      76.5  0.0361
# 3 2015  -1.39±5.3   -1        56.5  0.211 




# Additionnaly check first occurrence per year
week_first_occ <- samples %>% group_by(species,site,year) %>% summarise(week_first_occ=min(yweek))
# get ref week of arrival
week_first_occ_ref <- week_first_occ %>% group_by(species) %>% summarise(ref_week_first_occ = round(mean(week_first_occ),0))

# join the mean min week per species and the min week per year
lag_week_first_occ <- left_join(week_first_occ, week_first_occ_ref, by="species")
lag_week_first_occ$lag <- lag_week_first_occ$week_first_occ - lag_week_first_occ$ref_week_first_occ
lag_week_first_occ <- filter(lag_week_first_occ, !year == 2012)


descriptive_stats <- lag_week_first_occ %>%
  group_by(year) %>%
  summarise(stats=str_c(round(mean(lag), 2), "±", round(sd(lag),1)), median=median(lag))

suppressWarnings (test <- lag_week_first_occ %>% group_by(year) %>% do(glance(wilcox.test(.$lag,mu = 0,alternative = "two.sided"))))
left_join(descriptive_stats, select(test, statistic, p.value), by="year") %>% ungroup()
#   year  stats     median statistic p.value
#   <fct> <chr>      <dbl>     <dbl>   <dbl>
# 1 2013  2±8          1.5      93.5  0.434 
# 2 2014  -0.86±5.6   -1        69    0.0359
# 3 2015  -1.48±5.9   -1        75    0.269 




# Compute differences among years
diff2014_2013 <- filter(week_first_occ, year == 2014)$ref_week_first_occ - filter(week_first_occ, year == 2013)$ref_week_first_occ
diff2015_2013 <- filter(week_first_occ, year == 2015)$ref_week_first_occ - filter(week_first_occ, year == 2013, !species=="Scorpaena porcus")$ref_week_first_occ
diff2015_2014 <- filter(week_first_occ, year == 2015)$ref_week_first_occ - filter(week_first_occ, year == 2014, !species=="Scorpaena porcus")$ref_week_first_occ
# S. porcus was not caught in 2015 and so it shambled the difference in the timing of first arrival 

# Print differences between years
data.frame(year=c("diff2014_2013", "diff2015_2013", "diff2015_2014"), diff=c(
  str_c(round_any(mean(diff2014_2013), 0.01), "±", round_any(sd(diff2014_2013), 0.01)),
  str_c(round_any(mean(diff2015_2013), 0.01), "±", round_any(sd(diff2015_2013), 0.01)),
  diff2015_2014 = str_c(round_any(mean(diff2015_2014), 0.01), "±", round_any(sd(diff2015_2014), 0.01))))


# Check if significant lower in the next year
suppressWarnings (data.frame(rbind(data.frame(first_occ = diff2014_2013), data.frame(first_occ = diff2015_2014), data.frame(first_occ = diff2015_2013)), years = c(rep("2014-2013", 18), rep("2015-2014", 17), rep("2015-2013", 14))) %>% group_by(years) %>% do(glance(wilcox.test(.$first_occ, mu = 0, alternative = "less"))))
#       years statistic p.value                                               method alternative
# 1 2014-2013      66   0.317   Wilcoxon signed rank test with continuity correction less       
# 2 2015-2013      18   0.00171 Wilcoxon signed rank test with continuity correction less       
# 3 2015-2014      43.5 0.0611  Wilcoxon signed rank test with continuity correction less 
# No significant difference in the timing of first arrival except between 2015-2013


# Same but two-sided
suppressWarnings (data.frame(rbind(data.frame(first_occ = diff2014_2013), data.frame(first_occ = diff2015_2014), data.frame(first_occ = diff2015_2013)), years = c(rep("2014-2013", 18), rep("2015-2014", 17), rep("2015-2013", 14))) %>% group_by(years) %>% do(glance(wilcox.test(.$first_occ,mu = 0,alternative = "two"))))
#   years     statistic p.value method                                               alternative
# 1 2014-2013      66   0.635   Wilcoxon signed rank test with continuity correction two.sided  
# 2 2015-2013      18   0.00342 Wilcoxon signed rank test with continuity correction two.sided  
# 3 2015-2014      43.5 0.122   Wilcoxon signed rank test with continuity correction two.sided 

# Although providing pretty similar results, testing for difference in the mean timing of arrival is more relevant and provides "better"" results (here the p is 0.06, which is tricky to address in the manuscript)




#--- Time arrival per year and topography

samples <- summer
samples <- rbind(summer, winter)

# join with topography
samples <- left_join(samples, select(sites, site, rhone, topography))

ref_detailed <- samples %>% group_by(species,site,topography,year) %>% summarise(ref_week_det=round(weighted.mean(yweek, cpue),0))
ref_sp <- ref_detailed %>% group_by(species, topography) %>% summarise(ref_week_sp = round(mean(ref_week_det),0))

#Join the two data.frame to comute lag time for each site
lag <- left_join(ref_detailed, ref_sp, by=c("species", "topography"))
lag$lag_week <- lag$ref_week_det - lag$ref_week_sp


#  Remove 2012 because not full
lag <- filter(lag, !year == 2012)

ggplot(lag) +
  geom_histogram(aes(x=lag_week),binwidth=1) +
  geom_density(aes(x=lag_week, y=..count..)) +
  facet_grid(year~topography)+geom_vline(xintercept = 0,linetype="dashed") + xlab("")

#Compute mean and median lag
descriptive_stats <- lag %>%
  group_by(year, topography) %>%
  summarise(stats=str_c(round(mean(lag_week),2), "±", round(sd(lag_week),1)), median=median(lag_week))


suppressWarnings (test <- lag %>% group_by(year, topography) %>% do(glance(wilcox.test(.$lag_week,mu = 0,alternative = "two.sided"))))


# make a nice table with both
left_join(descriptive_stats, select(test, statistic, p.value), by=c("year", "topography")) %>% ungroup()
# # A tibble: 6 × 6
#     year   topography     stats median statistic p.value
# 1 2013  Gulf of Lion 2.42±5.8       1      502. 0.00197  
# 2 2013  Ligurian sea -0.64±5.2     -1      493  0.232    
# 3 2014  Gulf of Lion -0.29±7.7     -1      344  0.180    
# 4 2014  Ligurian sea -0.41±6       -1      734  0.182    
# 5 2015  Gulf of Lion -3.88±5.5     -3       53  0.0000466
# 6 2015  Ligurian sea -0.24±5.9     -1      116. 0.498   
 # Considering or not the winter species does not influence the results (same site*year are significant)

# summer only
#   year  topography   stats     median statistic   p.value
#   <fct> <fct>        <chr>      <dbl>     <dbl>     <dbl>
# 1 2013  Gulf of Lion 1.62±3.7       1     352.  0.00318  
# 2 2013  Ligurian sea -0.72±4.2      0     438   0.504    
# 3 2014  Gulf of Lion -0.7±4.5      -1     254.  0.144    
# 4 2014  Ligurian sea -0.2±4.4      -1     598.  0.296    
# 5 2015  Gulf of Lion -2.56±2.4     -2      24   0.0000688
# 6 2015  Ligurian sea -1.05±4.2     -1      68.5 0.284   



#--- Time arrival per year and rhone side


#Compute the mean arrival week by species/sites/year
ref_detailed <- samples %>% group_by(species,site,rhone,year) %>% summarise(ref_week_det=round(weighted.mean(yweek, cpue),0))
ref_sp <- ref_detailed %>% group_by(species, rhone) %>% summarise(ref_week_sp = round(mean(ref_week_det),0))


#Join the two data.frame to comute lag time for each site
lag <- left_join(ref_detailed, ref_sp, by=c("species", "rhone"))
lag$lag_week <- lag$ref_week_det - lag$ref_week_sp


#  Remove 2012 because not full
lag <- filter(lag, !year == 2012)

ggplot(lag) +
  geom_histogram(aes(x=lag_week),binwidth=1) +
  geom_density(aes(x=lag_week, y=..count..)) +
  facet_grid(year~rhone)+geom_vline(xintercept = 0,linetype="dashed") + xlab("")


#Compute mean and median lag
descriptive_stats <- lag %>%
  group_by(year, rhone) %>%
  summarise(stats=str_c(round(mean(lag_week),2), "±", round(sd(lag_week),1)), median=median(lag_week))


suppressWarnings (test <- lag %>% group_by(year, rhone) %>% do(glance(wilcox.test(.$lag_week,mu = 0,alternative = "two.sided"))))


# make a nice table with both
left_join(descriptive_stats, select(test, statistic, p.value), by=c("year", "rhone")) %>% ungroup()
#     year         rhone     stats median statistic p.value
# 1 2013  West of Rhone 1.07±4       0        165  0.215  
# 2 2013  East of Rhone 0.51±6.2     0        852. 0.489  
# 3 2014  West of Rhone -0.93±8.1   -2         91  0.0938 
# 4 2014  East of Rhone -0.25±6.1   -1       1197  0.228  
# 5 2015  West of Rhone -13.5±7.8  -13.5        0  0.5    
# 6 2015  East of Rhone -1.66±5.4   -1        315  0.00492




```

```{r, fig.height=9}


###--- Time arrival per year and site

#Compute the mean arrival week by species/sites/year
ref_detailed <- samples %>% group_by(species,site,year) %>% summarise(ref_week_det=round(weighted.mean(yweek, cpue),0))
ref_sp <- ref_detailed %>% group_by(species, site) %>% summarise(ref_week_sp = round(mean(ref_week_det),0))


#Join the two data.frame to comute lag time for each site
lag <- left_join(ref_detailed, ref_sp, by=c("species", "site"))
lag$lag_week <- lag$ref_week_det - lag$ref_week_sp


#  Remove 2012 because not full
lag <- filter(lag, !year == 2012)

ggplot(lag) +
  geom_histogram(aes(x=lag_week),binwidth=1) +
  geom_density(aes(x=lag_week, y=..count..)) +
  facet_grid(site~year)+geom_vline(xintercept = 0,linetype="dashed") + xlab("")

#Compute mean and median lag
descriptive_stats <- lag %>%
  group_by(site, year) %>%
  summarise(stats=str_c(round(mean(lag_week),2), "±", round(sd(lag_week),1)), median=median(lag_week))


suppressWarnings (test <- lag %>% group_by(site, year) %>% do(glance(wilcox.test(.$lag_week,mu = 0,alternative = "two.sided"))))


# make a nice table with both
data.frame(left_join(descriptive_stats, select(test, statistic, p.value), by=c("site", "year")) %>% ungroup())
# # A tibble: 26 × 6
#      year          site     stats median statistic p.value
#    <fctr>        <fctr>     <chr>  <dbl>     <dbl>   <dbl>
# 1    2013       Leucate      -2±8    0.0      25.0   0.837
# 2    2013  Port Vendres     0±1.3    0.0       5.0   1.000
# 3    2013          Agde     0±1.2    0.0       5.0   1.000
# 4    2013     Marseille     0±4.7    0.0       5.0   1.000
# 5    2013     Port-Cros   0.1±1.1    0.0       8.5   0.890
# 6    2013  Villefranche -0.56±3.6    0.0      13.5   1.000
# 7    2013     Bonifacio -0.57±1.5    0.0       2.5   0.424
# 8    2013 Saint Florent -1.31±2.6    0.0       6.5   0.121
# 9    2013        Bastia  -0.2±6.8    0.0      18.0   0.632
# 10   2014       Leucate -2.21±4.4   -1.5      11.5   0.033


```
